@@ -1,28 +1,20 @@
 import 'package:polygonid_flutter_sdk/assets/onchain_non_merkelized_issuer_base.g.dart';
 import 'package:polygonid_flutter_sdk/assets/get_issuer_id_interface.g.dart';
 import 'package:polygonid_flutter_sdk/common/domain/domain_constants.dart';
-import 'package:polygonid_flutter_sdk/common/domain/entities/chain_config_entity.dart';
-import 'package:polygonid_flutter_sdk/common/domain/entities/env_entity.dart';
 import 'package:polygonid_flutter_sdk/common/domain/error_exception.dart';
 import 'package:polygonid_flutter_sdk/common/domain/use_cases/get_env_use_case.dart';
 import 'package:polygonid_flutter_sdk/common/domain/use_cases/get_selected_chain_use_case.dart';
 import 'package:polygonid_flutter_sdk/common/infrastructure/stacktrace_stream_manager.dart';
-import 'package:polygonid_flutter_sdk/credential/domain/use_cases/cache_credential_use_case.dart';
-import 'package:polygonid_flutter_sdk/iden3comm/domain/entities/credential/request/base.dart';
-import 'package:polygonid_flutter_sdk/iden3comm/domain/entities/credential/request/onchain_offer_iden3_message_entity.dart';
-import 'package:polygonid_flutter_sdk/iden3comm/domain/entities/interaction/interaction_entity.dart';
 import 'package:polygonid_flutter_sdk/iden3comm/domain/exceptions/iden3comm_exceptions.dart';
 import 'package:polygonid_flutter_sdk/iden3comm/domain/repositories/did_profile_info_repository.dart';
 import 'package:polygonid_flutter_sdk/iden3comm/domain/use_cases/check_profile_and_did_current_env.dart';
 import 'package:polygonid_flutter_sdk/iden3comm/domain/use_cases/fetch_onchain_claim_use_case.dart';
 import 'package:polygonid_flutter_sdk/identity/data/data_sources/local_contract_files_data_source.dart';
 import 'package:polygonid_flutter_sdk/identity/domain/repositories/identity_repository.dart';
-import 'package:polygonid_flutter_sdk/identity/domain/use_cases/get_did_identifier_use_case.dart';
 import 'package:polygonid_flutter_sdk/constants.dart';
 import 'package:polygonid_flutter_sdk/common/domain/domain_logger.dart';
 import 'package:polygonid_flutter_sdk/common/domain/use_case.dart';
 import 'package:polygonid_flutter_sdk/credential/domain/entities/claim_entity.dart';
-import 'package:polygonid_flutter_sdk/credential/domain/use_cases/save_claims_use_case.dart';
 import 'package:polygonid_flutter_sdk/identity/domain/use_cases/get_did_use_case.dart';
 import 'package:polygonid_flutter_sdk/identity/domain/use_cases/get_public_keys_use_case.dart';
 import 'package:polygonid_flutter_sdk/proof/data/dtos/atomic_query_inputs_config_param.dart';
@@ -33,14 +25,14 @@ import 'package:web3dart/web3dart.dart';
 class FetchOnchainClaimsParam {
   final String contractAddress;
   final String genesisDid;
-  final BigInt profileNonce;
+  final BigInt? profileNonce;
   final String privateKey;
   final String? chainId;
 
   FetchOnchainClaimsParam({
     required this.contractAddress,
     required this.genesisDid,
-    required this.profileNonce,
+    this.profileNonce,
     required this.privateKey,
     this.chainId,
   });
@@ -53,10 +45,7 @@ class FetchOnchainClaimsUseCase
       _checkProfileAndDidCurrentEnvUseCase;
   final GetEnvUseCase _getEnvUseCase;
   final GetSelectedChainUseCase _getSelectedChainUseCase;
-  final GetDidIdentifierUseCase _getDidIdentifierUseCase;
   final GetDidUseCase _getDidUseCase;
-  final SaveClaimsUseCase _saveClaimsUseCase;
-  final CacheCredentialUseCase _cacheCredentialUseCase;
   final GetPublicKeyUseCase _getPublicKeyUseCase;
   final LocalContractFilesDataSource _localContractFilesDataSource;
   final IdentityRepository _identityRepository;
@@ -68,10 +57,7 @@ class FetchOnchainClaimsUseCase
     this._checkProfileAndDidCurrentEnvUseCase,
     this._getEnvUseCase,
     this._getSelectedChainUseCase,
-    this._getDidIdentifierUseCase,
     this._getDidUseCase,
-    this._saveClaimsUseCase,
-    this._cacheCredentialUseCase,
     this._getPublicKeyUseCase,
     this._localContractFilesDataSource,
     this._identityRepository,
@@ -175,7 +161,7 @@ class FetchOnchainClaimsUseCase
     final issuerIdInt = await getIssuerId.getId();
     final issuerDid = (await _identityRepository.describeId(
       id: issuerIdInt,
-      config: ConfigParam.fromEnv(env),
+      config: env.config,
     ))
         .did;
 
@@ -187,7 +173,9 @@ class FetchOnchainClaimsUseCase
     );
 
     BigInt nonce;
-    if (info.containsKey("privateProfileNonce")) {
+    if (param.profileNonce != null) {
+      nonce = param.profileNonce!;
+    } else if (info.containsKey("privateProfileNonce")) {
       nonce = BigInt.parse(info["privateProfileNonce"] as String);
     } else {
       nonce = GENESIS_PROFILE_NONCE;
@@ -203,7 +191,14 @@ class FetchOnchainClaimsUseCase
       ),
     );
 
-    final didEntity = await _getDidUseCase.execute(param: did);
+    // TODO OR
+
+    final profileDid = _identityRepository.getPrivateProfileForGenesisDid(
+      genesisDid: param.genesisDid,
+      profileNonce: nonce,
+    );
+
+    final didEntity = await _getDidUseCase.execute(param: profileDid);
     final userId =
         await _identityRepository.convertIdToBigInt(id: didEntity.identifier);
 
