@@ -9,14 +9,11 @@ import '../../../common/utils/hex_utils.dart';
 import '../../../common/utils/uint8_list_utils.dart';
 import 'bjj.dart';
 
-import 'package:polygonid_flutter_sdk/common/kms/keys/public_key.dart';
-import 'package:polygonid_flutter_sdk/common/kms/keys/private_key.dart';
-
-final bjjLib = BabyjubjubLib();
+typedef Point = ({BigInt x, BigInt y});
 
 /// Class representing EdDSA Baby Jub signature
 class BjjSignature {
-  late List<BigInt> r8;
+  late Point r8;
   late BigInt s;
 
   /// Create a Signature with the R8 point and S scalar
@@ -33,86 +30,64 @@ class BjjSignature {
     }
     BabyjubjubLib bjjLib = BabyjubjubLib();
 
-    final XYSublist = buf.sublist(0, 32);
-    // unpackPoint is used to unpack R8 X and Y
-    final List<String>? unpackedPoint =
-        bjjLib.unpackPoint(HEX.encode(XYSublist.toList()));
+    final pointsSublist = buf.sublist(0, 32);
+    // uncompressPoint is used to unpack R8 X and Y
+    final unpackedPoint =
+        bjjLib.uncompressPoint(HEX.encode(pointsSublist.toList()));
 
-    BigInt? x = BigInt.tryParse(unpackedPoint![0], radix: 10);
-    BigInt? y = BigInt.tryParse(unpackedPoint[1], radix: 10);
-    List<BigInt> r8 = [];
-    r8.add(x!);
-    r8.add(y!);
+    final x = BigInt.parse(unpackedPoint[0]);
+    final y = BigInt.parse(unpackedPoint[1]);
 
     // S is decoded manually
     BigInt s = Uint8ArrayUtils.beBuff2int(
         Uint8List.fromList(buf.sublist(32, 64).reversed.toList()));
-    return BjjSignature(r8, s);
+    return BjjSignature((x: x, y: y), s);
+  }
+
+  /// Compress the Signature
+  /// @returns {Uint8List} - signature compressed into a buffer
+  Uint8List compress() {
+    final r8 = HexUtils.hexToBytes(
+      BabyjubjubLib().compressPoint(this.r8.x.toString(), this.r8.y.toString()),
+    );
+    final s = Uint8ArrayUtils.beInt2Buff(this.s, 32);
+    return Uint8List.fromList([...r8, ...s]);
   }
 }
 
 /// Class representing a EdDSA Baby Jub public key
 class BjjPublicKey extends PublicKey {
-  final List<BigInt> p;
+  final String hex;
+  final Point p;
 
   /// Create a PublicKey from a curve point p
+  /// @param {String} hex - hex representation
   /// @param {List[BigInt]} p - curve point
-  BjjPublicKey(this.p)
-      : super(
-          hex: bjjLib.packPoint(
-            p[0].toString(),
-            p[1].toString(),
-          ),
-        );
+  BjjPublicKey._(this.hex, this.p);
 
-  KeyType get keyType => KeyType.BabyJubJub;
-
-  /// Create a PublicKey from a bigInt compressed pubKey
-  ///
-  /// @param {BigInt} compressedBigInt - compressed public key in a bigInt
-  ///
-  /// @returns {PublicKey} public key class
-  factory BjjPublicKey.newFromCompressed(BigInt compressedBigInt) {
-    final Uint8List compressedBuffLE =
-        Uint8ArrayUtils.leInt2Buff(compressedBigInt, 32);
-    if (compressedBuffLE.length != 32) {
-      throw ArgumentError('buf must be 32 bytes');
-    }
-    final p =
-        bjjLib.unpackPoint(Uint8ArrayUtils.uint8ListToString(compressedBuffLE));
-    if (p == null) {
-      throw ArgumentError('unpackPoint failed');
-    }
+  factory BjjPublicKey.hex(String hex) {
+    BabyjubjubLib bjjLib = BabyjubjubLib();
+    final p = bjjLib.uncompressPoint(hex);
     BigInt x = BigInt.parse(p[0]);
     BigInt y = BigInt.parse(p[1]);
-    List<BigInt> point = [];
-    point.add(x);
-    point.add(y);
-    return BjjPublicKey(point);
+
+    return PublicKey._(hex, (x: x, y: y));
   }
 
   /// Compress the PublicKey
   /// @returns {Uint8List} - point compressed into a buffer
   Uint8List compress() {
+    BabyjubjubLib bjjLib = BabyjubjubLib();
     return HexUtils.hexToBytes(
-      bjjLib.packPoint(
-        p[0].toString(),
-        p[1].toString(),
-      ),
+      bjjLib.compressPoint(p.x.toString(), p.y.toString()),
     );
   }
 
-  bool verify(String messageHash, BjjSignature signature) {
-    List<int> pointList = [];
-    pointList.add(p[0].toInt());
-    pointList.add(p[1].toInt());
-    List<int> sigList = [];
-    sigList.add(signature.r8[0].toInt());
-    sigList.add(signature.r8[1].toInt());
-    sigList.add(signature.s.toInt());
+  bool verify(String messageHash, Signature signature) {
+    BabyjubjubLib bjjLib = BabyjubjubLib();
     return bjjLib.verifyPoseidon(
-      Uint8ArrayUtils.uint8ListToString(Uint8List.fromList(pointList)),
-      Uint8ArrayUtils.uint8ListToString(Uint8List.fromList(sigList)),
+      hex,
+      Uint8ArrayUtils.uint8ListToString(signature.compress()),
       messageHash,
     );
   }
@@ -133,32 +108,20 @@ class BjjPrivateKey extends PrivateKey {
 
   /// Retrieve PublicKey of the PrivateKey
   /// @returns {PublicKey} PublicKey derived from PrivateKey
-  @override
   BjjPublicKey publicKey() {
-    String resultString = bjjLib.prv2pub(HexUtils.bytesToHex(sk));
-    final stringList = resultString.split(",");
-    stringList[0] = stringList[0].replaceAll("Fr(", "");
-    stringList[0] = stringList[0].replaceAll(")", "");
-    stringList[1] = stringList[1].replaceAll("Fr(", "");
-    stringList[1] = stringList[1].replaceAll(")", "");
-    BigInt x = HexUtils.hexToInt(stringList[0]);
-    BigInt y = HexUtils.hexToInt(stringList[1]);
-    List<BigInt> p = [];
-    p.add(x);
-    p.add(y);
-    return BjjPublicKey(p);
+    BabyjubjubLib bjjLib = BabyjubjubLib();
+    String publicKeyHex = bjjLib.prv2pub(HexUtils.bytesToHex(sk));
+    return BjjPublicKey.hex(publicKeyHex);
   }
 
-  @override
-  Uint8List sign(Uint8List message) {
-    final messageHashBytes = bytesToHex(message);
-
-    final messageHashBigInt = BigInt.parse(messageHashBytes, radix: 16);
-
+  String sign(BigInt messageHash) {
+    BabyjubjubLib bjjLib = BabyjubjubLib();
     String signature = bjjLib.signPoseidon(
       HexUtils.bytesToHex(sk),
-      messageHashBigInt.toString(),
+      messageHash.toString(),
     );
+    return signature;
+  }
 
     return Uint8ArrayUtils.uint8ListfromString(signature);
   }
