@@ -1,7 +1,7 @@
 // AUTO GENERATED FILE, DO NOT EDIT.
 //
 // Generated by `package:ffigen`.
-// ignore_for_file: type=lint
+// ignore_for_file: type=lint, unused_element, unused_field
 import 'dart:ffi' as ffi;
 
 /// Bindings to `ios/Classes/libpolygonid.h`.
@@ -188,13 +188,13 @@ class NativePolygonIdCoreLib {
       _waitpidPtr.asFunction<int Function(int, ffi.Pointer<ffi.Int>, int)>();
 
   int waitid(
-    int arg0,
-    int arg1,
+    idtype_t arg0,
+    Dart__uint32_t arg1,
     ffi.Pointer<siginfo_t> arg2,
     int arg3,
   ) {
     return _waitid(
-      arg0,
+      arg0.value,
       arg1,
       arg2,
       arg3,
@@ -203,8 +203,8 @@ class NativePolygonIdCoreLib {
 
   late final _waitidPtr = _lookup<
       ffi.NativeFunction<
-          ffi.Int Function(
-              ffi.Int32, id_t, ffi.Pointer<siginfo_t>, ffi.Int)>>('waitid');
+          ffi.Int Function(ffi.UnsignedInt, id_t, ffi.Pointer<siginfo_t>,
+              ffi.Int)>>('waitid');
   late final _waitid = _waitidPtr
       .asFunction<int Function(int, int, ffi.Pointer<siginfo_t>, int)>();
 
@@ -606,10 +606,10 @@ class NativePolygonIdCoreLib {
       .asFunction<ffi.Pointer<ffi.Void> Function(ffi.Pointer<ffi.Void>, int)>();
 
   ffi.Pointer<ffi.Void> valloc(
-    int arg0,
+    int __size,
   ) {
     return _valloc(
-      arg0,
+      __size,
     );
   }
 
@@ -3370,29 +3370,6 @@ class NativePolygonIdCoreLib {
       _lookup<ffi.NativeFunction<ffi.Int Function(ffi.LongLong)>>('flsll');
   late final _flsll = _flsllPtr.asFunction<int Function(int)>();
 
-  int PLGNAuthV2InputsMarshal(
-    ffi.Pointer<ffi.Pointer<ffi.Char>> jsonResponse,
-    ffi.Pointer<ffi.Char> in1,
-    ffi.Pointer<ffi.Pointer<PLGNStatus>> status,
-  ) {
-    return _PLGNAuthV2InputsMarshal(
-      jsonResponse,
-      in1,
-      status,
-    );
-  }
-
-  late final _PLGNAuthV2InputsMarshalPtr = _lookup<
-          ffi.NativeFunction<
-              GoUint8 Function(
-                  ffi.Pointer<ffi.Pointer<ffi.Char>>,
-                  ffi.Pointer<ffi.Char>,
-                  ffi.Pointer<ffi.Pointer<PLGNStatus>>)>>(
-      'PLGNAuthV2InputsMarshal');
-  late final _PLGNAuthV2InputsMarshal = _PLGNAuthV2InputsMarshalPtr.asFunction<
-      int Function(ffi.Pointer<ffi.Pointer<ffi.Char>>, ffi.Pointer<ffi.Char>,
-          ffi.Pointer<ffi.Pointer<PLGNStatus>>)>();
-
   int PLGNCalculateGenesisID(
     ffi.Pointer<ffi.Pointer<ffi.Char>> jsonResponse,
     ffi.Pointer<ffi.Char> in1,
@@ -3806,6 +3783,31 @@ class NativePolygonIdCoreLib {
               ffi.Pointer<ffi.Char>,
               ffi.Pointer<ffi.Pointer<PLGNStatus>>)>();
 
+  int PLGNAGenerateInputs(
+    ffi.Pointer<ffi.Pointer<ffi.Char>> jsonResponse,
+    ffi.Pointer<ffi.Char> in1,
+    ffi.Pointer<ffi.Char> cfg,
+    ffi.Pointer<ffi.Pointer<PLGNStatus>> status,
+  ) {
+    return _PLGNAGenerateInputs(
+      jsonResponse,
+      in1,
+      cfg,
+      status,
+    );
+  }
+
+  late final _PLGNAGenerateInputsPtr = _lookup<
+      ffi.NativeFunction<
+          GoUint8 Function(
+              ffi.Pointer<ffi.Pointer<ffi.Char>>,
+              ffi.Pointer<ffi.Char>,
+              ffi.Pointer<ffi.Char>,
+              ffi.Pointer<ffi.Pointer<PLGNStatus>>)>>('PLGNAGenerateInputs');
+  late final _PLGNAGenerateInputs = _PLGNAGenerateInputsPtr.asFunction<
+      int Function(ffi.Pointer<ffi.Pointer<ffi.Char>>, ffi.Pointer<ffi.Char>,
+          ffi.Pointer<ffi.Char>, ffi.Pointer<ffi.Pointer<PLGNStatus>>)>();
+
   void PLGNFreeStatus(
     ffi.Pointer<PLGNStatus> status,
   ) {
@@ -3834,6 +3836,24 @@ class NativePolygonIdCoreLib {
               ffi.Pointer<ffi.Pointer<PLGNStatus>>)>>('PLGNCleanCache');
   late final _PLGNCleanCache = _PLGNCleanCachePtr.asFunction<
       int Function(ffi.Pointer<ffi.Pointer<PLGNStatus>>)>();
+
+  int PLGNCleanCache2(
+    ffi.Pointer<ffi.Char> cfg,
+    ffi.Pointer<ffi.Pointer<PLGNStatus>> status,
+  ) {
+    return _PLGNCleanCache2(
+      cfg,
+      status,
+    );
+  }
+
+  late final _PLGNCleanCache2Ptr = _lookup<
+      ffi.NativeFunction<
+          GoUint8 Function(ffi.Pointer<ffi.Char>,
+              ffi.Pointer<ffi.Pointer<PLGNStatus>>)>>('PLGNCleanCache2');
+  late final _PLGNCleanCache2 = _PLGNCleanCache2Ptr.asFunction<
+      int Function(
+          ffi.Pointer<ffi.Char>, ffi.Pointer<ffi.Pointer<PLGNStatus>>)>();
 
   int PLGNCacheCredentials(
     ffi.Pointer<ffi.Char> in1,
@@ -3909,9 +3929,159 @@ class NativePolygonIdCoreLib {
   late final _PLGNDescribeID = _PLGNDescribeIDPtr.asFunction<
       int Function(ffi.Pointer<ffi.Pointer<ffi.Char>>, ffi.Pointer<ffi.Char>,
           ffi.Pointer<ffi.Char>, ffi.Pointer<ffi.Pointer<PLGNStatus>>)>();
+
+  int PLGNBabyJubJubSignPoseidon(
+    ffi.Pointer<ffi.Pointer<ffi.Char>> jsonResponse,
+    ffi.Pointer<ffi.Char> in1,
+    ffi.Pointer<ffi.Char> cfg,
+    ffi.Pointer<ffi.Pointer<PLGNStatus>> status,
+  ) {
+    return _PLGNBabyJubJubSignPoseidon(
+      jsonResponse,
+      in1,
+      cfg,
+      status,
+    );
+  }
+
+  late final _PLGNBabyJubJubSignPoseidonPtr = _lookup<
+          ffi.NativeFunction<
+              GoUint8 Function(
+                  ffi.Pointer<ffi.Pointer<ffi.Char>>,
+                  ffi.Pointer<ffi.Char>,
+                  ffi.Pointer<ffi.Char>,
+                  ffi.Pointer<ffi.Pointer<PLGNStatus>>)>>(
+      'PLGNBabyJubJubSignPoseidon');
+  late final _PLGNBabyJubJubSignPoseidon =
+      _PLGNBabyJubJubSignPoseidonPtr.asFunction<
+          int Function(
+              ffi.Pointer<ffi.Pointer<ffi.Char>>,
+              ffi.Pointer<ffi.Char>,
+              ffi.Pointer<ffi.Char>,
+              ffi.Pointer<ffi.Pointer<PLGNStatus>>)>();
+
+  int PLGNBabyJubJubVerifyPoseidon(
+    ffi.Pointer<ffi.Pointer<ffi.Char>> jsonResponse,
+    ffi.Pointer<ffi.Char> in1,
+    ffi.Pointer<ffi.Char> cfg,
+    ffi.Pointer<ffi.Pointer<PLGNStatus>> status,
+  ) {
+    return _PLGNBabyJubJubVerifyPoseidon(
+      jsonResponse,
+      in1,
+      cfg,
+      status,
+    );
+  }
+
+  late final _PLGNBabyJubJubVerifyPoseidonPtr = _lookup<
+          ffi.NativeFunction<
+              GoUint8 Function(
+                  ffi.Pointer<ffi.Pointer<ffi.Char>>,
+                  ffi.Pointer<ffi.Char>,
+                  ffi.Pointer<ffi.Char>,
+                  ffi.Pointer<ffi.Pointer<PLGNStatus>>)>>(
+      'PLGNBabyJubJubVerifyPoseidon');
+  late final _PLGNBabyJubJubVerifyPoseidon =
+      _PLGNBabyJubJubVerifyPoseidonPtr.asFunction<
+          int Function(
+              ffi.Pointer<ffi.Pointer<ffi.Char>>,
+              ffi.Pointer<ffi.Char>,
+              ffi.Pointer<ffi.Char>,
+              ffi.Pointer<ffi.Pointer<PLGNStatus>>)>();
+
+  int PLGNBabyJubJubPrivate2Public(
+    ffi.Pointer<ffi.Pointer<ffi.Char>> jsonResponse,
+    ffi.Pointer<ffi.Char> in1,
+    ffi.Pointer<ffi.Char> cfg,
+    ffi.Pointer<ffi.Pointer<PLGNStatus>> status,
+  ) {
+    return _PLGNBabyJubJubPrivate2Public(
+      jsonResponse,
+      in1,
+      cfg,
+      status,
+    );
+  }
+
+  late final _PLGNBabyJubJubPrivate2PublicPtr = _lookup<
+          ffi.NativeFunction<
+              GoUint8 Function(
+                  ffi.Pointer<ffi.Pointer<ffi.Char>>,
+                  ffi.Pointer<ffi.Char>,
+                  ffi.Pointer<ffi.Char>,
+                  ffi.Pointer<ffi.Pointer<PLGNStatus>>)>>(
+      'PLGNBabyJubJubPrivate2Public');
+  late final _PLGNBabyJubJubPrivate2Public =
+      _PLGNBabyJubJubPrivate2PublicPtr.asFunction<
+          int Function(
+              ffi.Pointer<ffi.Pointer<ffi.Char>>,
+              ffi.Pointer<ffi.Char>,
+              ffi.Pointer<ffi.Char>,
+              ffi.Pointer<ffi.Pointer<PLGNStatus>>)>();
+
+  int PLGNBabyJubJubPublicUncompress(
+    ffi.Pointer<ffi.Pointer<ffi.Char>> jsonResponse,
+    ffi.Pointer<ffi.Char> in1,
+    ffi.Pointer<ffi.Char> cfg,
+    ffi.Pointer<ffi.Pointer<PLGNStatus>> status,
+  ) {
+    return _PLGNBabyJubJubPublicUncompress(
+      jsonResponse,
+      in1,
+      cfg,
+      status,
+    );
+  }
+
+  late final _PLGNBabyJubJubPublicUncompressPtr = _lookup<
+          ffi.NativeFunction<
+              GoUint8 Function(
+                  ffi.Pointer<ffi.Pointer<ffi.Char>>,
+                  ffi.Pointer<ffi.Char>,
+                  ffi.Pointer<ffi.Char>,
+                  ffi.Pointer<ffi.Pointer<PLGNStatus>>)>>(
+      'PLGNBabyJubJubPublicUncompress');
+  late final _PLGNBabyJubJubPublicUncompress =
+      _PLGNBabyJubJubPublicUncompressPtr.asFunction<
+          int Function(
+              ffi.Pointer<ffi.Pointer<ffi.Char>>,
+              ffi.Pointer<ffi.Char>,
+              ffi.Pointer<ffi.Char>,
+              ffi.Pointer<ffi.Pointer<PLGNStatus>>)>();
+
+  int PLGNBabyJubJubPublicCompress(
+    ffi.Pointer<ffi.Pointer<ffi.Char>> jsonResponse,
+    ffi.Pointer<ffi.Char> in1,
+    ffi.Pointer<ffi.Char> cfg,
+    ffi.Pointer<ffi.Pointer<PLGNStatus>> status,
+  ) {
+    return _PLGNBabyJubJubPublicCompress(
+      jsonResponse,
+      in1,
+      cfg,
+      status,
+    );
+  }
+
+  late final _PLGNBabyJubJubPublicCompressPtr = _lookup<
+          ffi.NativeFunction<
+              GoUint8 Function(
+                  ffi.Pointer<ffi.Pointer<ffi.Char>>,
+                  ffi.Pointer<ffi.Char>,
+                  ffi.Pointer<ffi.Char>,
+                  ffi.Pointer<ffi.Pointer<PLGNStatus>>)>>(
+      'PLGNBabyJubJubPublicCompress');
+  late final _PLGNBabyJubJubPublicCompress =
+      _PLGNBabyJubJubPublicCompressPtr.asFunction<
+          int Function(
+              ffi.Pointer<ffi.Pointer<ffi.Char>>,
+              ffi.Pointer<ffi.Char>,
+              ffi.Pointer<ffi.Char>,
+              ffi.Pointer<ffi.Pointer<PLGNStatus>>)>();
 }
 
-class __mbstate_t extends ffi.Union {
+final class __mbstate_t extends ffi.Union {
   @ffi.Array.multi([128])
   external ffi.Array<ffi.Char> __mbstate8;
 
@@ -3919,7 +4089,7 @@ class __mbstate_t extends ffi.Union {
   external int _mbstateL;
 }
 
-class __darwin_pthread_handler_rec extends ffi.Struct {
+final class __darwin_pthread_handler_rec extends ffi.Struct {
   external ffi
       .Pointer<ffi.NativeFunction<ffi.Void Function(ffi.Pointer<ffi.Void>)>>
       __routine;
@@ -3929,7 +4099,7 @@ class __darwin_pthread_handler_rec extends ffi.Struct {
   external ffi.Pointer<__darwin_pthread_handler_rec> __next;
 }
 
-class _opaque_pthread_attr_t extends ffi.Struct {
+final class _opaque_pthread_attr_t extends ffi.Struct {
   @ffi.Long()
   external int __sig;
 
@@ -3937,7 +4107,7 @@ class _opaque_pthread_attr_t extends ffi.Struct {
   external ffi.Array<ffi.Char> __opaque;
 }
 
-class _opaque_pthread_cond_t extends ffi.Struct {
+final class _opaque_pthread_cond_t extends ffi.Struct {
   @ffi.Long()
   external int __sig;
 
@@ -3945,7 +4115,7 @@ class _opaque_pthread_cond_t extends ffi.Struct {
   external ffi.Array<ffi.Char> __opaque;
 }
 
-class _opaque_pthread_condattr_t extends ffi.Struct {
+final class _opaque_pthread_condattr_t extends ffi.Struct {
   @ffi.Long()
   external int __sig;
 
@@ -3953,7 +4123,7 @@ class _opaque_pthread_condattr_t extends ffi.Struct {
   external ffi.Array<ffi.Char> __opaque;
 }
 
-class _opaque_pthread_mutex_t extends ffi.Struct {
+final class _opaque_pthread_mutex_t extends ffi.Struct {
   @ffi.Long()
   external int __sig;
 
@@ -3961,7 +4131,7 @@ class _opaque_pthread_mutex_t extends ffi.Struct {
   external ffi.Array<ffi.Char> __opaque;
 }
 
-class _opaque_pthread_mutexattr_t extends ffi.Struct {
+final class _opaque_pthread_mutexattr_t extends ffi.Struct {
   @ffi.Long()
   external int __sig;
 
@@ -3969,7 +4139,7 @@ class _opaque_pthread_mutexattr_t extends ffi.Struct {
   external ffi.Array<ffi.Char> __opaque;
 }
 
-class _opaque_pthread_once_t extends ffi.Struct {
+final class _opaque_pthread_once_t extends ffi.Struct {
   @ffi.Long()
   external int __sig;
 
@@ -3977,7 +4147,7 @@ class _opaque_pthread_once_t extends ffi.Struct {
   external ffi.Array<ffi.Char> __opaque;
 }
 
-class _opaque_pthread_rwlock_t extends ffi.Struct {
+final class _opaque_pthread_rwlock_t extends ffi.Struct {
   @ffi.Long()
   external int __sig;
 
@@ -3985,7 +4155,7 @@ class _opaque_pthread_rwlock_t extends ffi.Struct {
   external ffi.Array<ffi.Char> __opaque;
 }
 
-class _opaque_pthread_rwlockattr_t extends ffi.Struct {
+final class _opaque_pthread_rwlockattr_t extends ffi.Struct {
   @ffi.Long()
   external int __sig;
 
@@ -3993,7 +4163,7 @@ class _opaque_pthread_rwlockattr_t extends ffi.Struct {
   external ffi.Array<ffi.Char> __opaque;
 }
 
-class _opaque_pthread_t extends ffi.Struct {
+final class _opaque_pthread_t extends ffi.Struct {
   @ffi.Long()
   external int __sig;
 
@@ -4003,7 +4173,7 @@ class _opaque_pthread_t extends ffi.Struct {
   external ffi.Array<ffi.Char> __opaque;
 }
 
-class _GoString_ extends ffi.Struct {
+final class _GoString_ extends ffi.Struct {
   external ffi.Pointer<ffi.Char> p;
 
   @ptrdiff_t()
@@ -4014,13 +4184,23 @@ typedef ptrdiff_t = __darwin_ptrdiff_t;
 typedef __darwin_ptrdiff_t = ffi.Long;
 typedef Dart__darwin_ptrdiff_t = int;
 
-abstract class idtype_t {
-  static const int P_ALL = 0;
-  static const int P_PID = 1;
-  static const int P_PGID = 2;
+enum idtype_t {
+  P_ALL(0),
+  P_PID(1),
+  P_PGID(2);
+
+  final int value;
+  const idtype_t(this.value);
+
+  static idtype_t fromValue(int value) => switch (value) {
+        0 => P_ALL,
+        1 => P_PID,
+        2 => P_PGID,
+        _ => throw ArgumentError("Unknown value for idtype_t: $value"),
+      };
 }
 
-class __darwin_arm_exception_state extends ffi.Struct {
+final class __darwin_arm_exception_state extends ffi.Struct {
   @__uint32_t()
   external int __exception;
 
@@ -4034,7 +4214,7 @@ class __darwin_arm_exception_state extends ffi.Struct {
 typedef __uint32_t = ffi.UnsignedInt;
 typedef Dart__uint32_t = int;
 
-class __darwin_arm_exception_state64 extends ffi.Struct {
+final class __darwin_arm_exception_state64 extends ffi.Struct {
   @__uint64_t()
   external int __far;
 
@@ -4048,7 +4228,7 @@ class __darwin_arm_exception_state64 extends ffi.Struct {
 typedef __uint64_t = ffi.UnsignedLongLong;
 typedef Dart__uint64_t = int;
 
-class __darwin_arm_exception_state64_v2 extends ffi.Struct {
+final class __darwin_arm_exception_state64_v2 extends ffi.Struct {
   @__uint64_t()
   external int __far;
 
@@ -4056,7 +4236,7 @@ class __darwin_arm_exception_state64_v2 extends ffi.Struct {
   external int __esr;
 }
 
-class __darwin_arm_thread_state extends ffi.Struct {
+final class __darwin_arm_thread_state extends ffi.Struct {
   @ffi.Array.multi([13])
   external ffi.Array<__uint32_t> __r;
 
@@ -4073,7 +4253,7 @@ class __darwin_arm_thread_state extends ffi.Struct {
   external int __cpsr;
 }
 
-class __darwin_arm_thread_state64 extends ffi.Struct {
+final class __darwin_arm_thread_state64 extends ffi.Struct {
   @ffi.Array.multi([29])
   external ffi.Array<__uint64_t> __x;
 
@@ -4096,7 +4276,7 @@ class __darwin_arm_thread_state64 extends ffi.Struct {
   external int __pad;
 }
 
-class __darwin_arm_vfp_state extends ffi.Struct {
+final class __darwin_arm_vfp_state extends ffi.Struct {
   @ffi.Array.multi([64])
   external ffi.Array<__uint32_t> __r;
 
@@ -4104,16 +4284,16 @@ class __darwin_arm_vfp_state extends ffi.Struct {
   external int __fpscr;
 }
 
-class __darwin_arm_neon_state64 extends ffi.Opaque {}
+final class __darwin_arm_neon_state64 extends ffi.Opaque {}
 
-class __darwin_arm_neon_state extends ffi.Opaque {}
+final class __darwin_arm_neon_state extends ffi.Opaque {}
 
-class __arm_pagein_state extends ffi.Struct {
+final class __arm_pagein_state extends ffi.Struct {
   @ffi.Int()
   external int __pagein_error;
 }
 
-class __arm_legacy_debug_state extends ffi.Struct {
+final class __arm_legacy_debug_state extends ffi.Struct {
   @ffi.Array.multi([16])
   external ffi.Array<__uint32_t> __bvr;
 
@@ -4127,7 +4307,7 @@ class __arm_legacy_debug_state extends ffi.Struct {
   external ffi.Array<__uint32_t> __wcr;
 }
 
-class __darwin_arm_debug_state32 extends ffi.Struct {
+final class __darwin_arm_debug_state32 extends ffi.Struct {
   @ffi.Array.multi([16])
   external ffi.Array<__uint32_t> __bvr;
 
@@ -4144,7 +4324,7 @@ class __darwin_arm_debug_state32 extends ffi.Struct {
   external int __mdscr_el1;
 }
 
-class __darwin_arm_debug_state64 extends ffi.Struct {
+final class __darwin_arm_debug_state64 extends ffi.Struct {
   @ffi.Array.multi([16])
   external ffi.Array<__uint64_t> __bvr;
 
@@ -4161,12 +4341,12 @@ class __darwin_arm_debug_state64 extends ffi.Struct {
   external int __mdscr_el1;
 }
 
-class __darwin_arm_cpmu_state64 extends ffi.Struct {
+final class __darwin_arm_cpmu_state64 extends ffi.Struct {
   @ffi.Array.multi([16])
   external ffi.Array<__uint64_t> __ctrs;
 }
 
-class __darwin_mcontext32 extends ffi.Struct {
+final class __darwin_mcontext32 extends ffi.Struct {
   external __darwin_arm_exception_state __es;
 
   external __darwin_arm_thread_state __ss;
@@ -4174,9 +4354,9 @@ class __darwin_mcontext32 extends ffi.Struct {
   external __darwin_arm_vfp_state __fs;
 }
 
-class __darwin_mcontext64 extends ffi.Opaque {}
+final class __darwin_mcontext64 extends ffi.Opaque {}
 
-class __darwin_sigaltstack extends ffi.Struct {
+final class __darwin_sigaltstack extends ffi.Struct {
   external ffi.Pointer<ffi.Void> ss_sp;
 
   @__darwin_size_t()
@@ -4189,7 +4369,7 @@ class __darwin_sigaltstack extends ffi.Struct {
 typedef __darwin_size_t = ffi.UnsignedLong;
 typedef Dart__darwin_size_t = int;
 
-class __darwin_ucontext extends ffi.Struct {
+final class __darwin_ucontext extends ffi.Struct {
   @ffi.Int()
   external int uc_onstack;
 
@@ -4208,14 +4388,14 @@ class __darwin_ucontext extends ffi.Struct {
 
 typedef __darwin_sigset_t = __uint32_t;
 
-class sigval extends ffi.Union {
+final class sigval extends ffi.Union {
   @ffi.Int()
   external int sival_int;
 
   external ffi.Pointer<ffi.Void> sival_ptr;
 }
 
-class sigevent extends ffi.Struct {
+final class sigevent extends ffi.Struct {
   @ffi.Int()
   external int sigev_notify;
 
@@ -4233,7 +4413,7 @@ class sigevent extends ffi.Struct {
 typedef pthread_attr_t = __darwin_pthread_attr_t;
 typedef __darwin_pthread_attr_t = _opaque_pthread_attr_t;
 
-class __siginfo extends ffi.Struct {
+final class __siginfo extends ffi.Struct {
   @ffi.Int()
   external int si_signo;
 
@@ -4270,7 +4450,7 @@ typedef Dart__int32_t = int;
 typedef uid_t = __darwin_uid_t;
 typedef __darwin_uid_t = __uint32_t;
 
-class __sigaction_u extends ffi.Union {
+final class __sigaction_u extends ffi.Union {
   external ffi.Pointer<ffi.NativeFunction<ffi.Void Function(ffi.Int)>>
       __sa_handler;
 
@@ -4281,7 +4461,7 @@ class __sigaction_u extends ffi.Union {
       __sa_sigaction;
 }
 
-class __sigaction extends ffi.Struct {
+final class __sigaction extends ffi.Struct {
   external __sigaction_u __sigaction_u1;
 
   external ffi.Pointer<
@@ -4299,7 +4479,7 @@ class __sigaction extends ffi.Struct {
 typedef siginfo_t = __siginfo;
 typedef sigset_t = __darwin_sigset_t;
 
-class sigaction extends ffi.Struct {
+final class sigaction extends ffi.Struct {
   external __sigaction_u __sigaction_u1;
 
   @sigset_t()
@@ -4309,7 +4489,7 @@ class sigaction extends ffi.Struct {
   external int sa_flags;
 }
 
-class sigvec extends ffi.Struct {
+final class sigvec extends ffi.Struct {
   external ffi.Pointer<ffi.NativeFunction<ffi.Void Function(ffi.Int)>>
       sv_handler;
 
@@ -4320,14 +4500,14 @@ class sigvec extends ffi.Struct {
   external int sv_flags;
 }
 
-class sigstack extends ffi.Struct {
+final class sigstack extends ffi.Struct {
   external ffi.Pointer<ffi.Char> ss_sp;
 
   @ffi.Int()
   external int ss_onstack;
 }
 
-class timeval extends ffi.Struct {
+final class timeval extends ffi.Struct {
   @__darwin_time_t()
   external int tv_sec;
 
@@ -4339,7 +4519,7 @@ typedef __darwin_time_t = ffi.Long;
 typedef Dart__darwin_time_t = int;
 typedef __darwin_suseconds_t = __int32_t;
 
-class rusage extends ffi.Struct {
+final class rusage extends ffi.Struct {
   external timeval ru_utime;
 
   external timeval ru_stime;
@@ -4387,7 +4567,7 @@ class rusage extends ffi.Struct {
   external int ru_nivcsw;
 }
 
-class rusage_info_v0 extends ffi.Struct {
+final class rusage_info_v0 extends ffi.Struct {
   @ffi.Array.multi([16])
   external ffi.Array<ffi.Uint8> ri_uuid;
 
@@ -4422,7 +4602,7 @@ class rusage_info_v0 extends ffi.Struct {
   external int ri_proc_exit_abstime;
 }
 
-class rusage_info_v1 extends ffi.Struct {
+final class rusage_info_v1 extends ffi.Struct {
   @ffi.Array.multi([16])
   external ffi.Array<ffi.Uint8> ri_uuid;
 
@@ -4475,7 +4655,7 @@ class rusage_info_v1 extends ffi.Struct {
   external int ri_child_elapsed_abstime;
 }
 
-class rusage_info_v2 extends ffi.Struct {
+final class rusage_info_v2 extends ffi.Struct {
   @ffi.Array.multi([16])
   external ffi.Array<ffi.Uint8> ri_uuid;
 
@@ -4534,7 +4714,7 @@ class rusage_info_v2 extends ffi.Struct {
   external int ri_diskio_byteswritten;
 }
 
-class rusage_info_v3 extends ffi.Struct {
+final class rusage_info_v3 extends ffi.Struct {
   @ffi.Array.multi([16])
   external ffi.Array<ffi.Uint8> ri_uuid;
 
@@ -4620,7 +4800,7 @@ class rusage_info_v3 extends ffi.Struct {
   external int ri_serviced_system_time;
 }
 
-class rusage_info_v4 extends ffi.Struct {
+final class rusage_info_v4 extends ffi.Struct {
   @ffi.Array.multi([16])
   external ffi.Array<ffi.Uint8> ri_uuid;
 
@@ -4730,7 +4910,7 @@ class rusage_info_v4 extends ffi.Struct {
   external int ri_runnable_time;
 }
 
-class rusage_info_v5 extends ffi.Struct {
+final class rusage_info_v5 extends ffi.Struct {
   @ffi.Array.multi([16])
   external ffi.Array<ffi.Uint8> ri_uuid;
 
@@ -4843,7 +5023,7 @@ class rusage_info_v5 extends ffi.Struct {
   external int ri_flags;
 }
 
-class rusage_info_v6 extends ffi.Struct {
+final class rusage_info_v6 extends ffi.Struct {
   @ffi.Array.multi([16])
   external ffi.Array<ffi.Uint8> ri_uuid;
 
@@ -4992,7 +5172,7 @@ class rusage_info_v6 extends ffi.Struct {
   external ffi.Array<ffi.Uint64> ri_reserved;
 }
 
-class rlimit extends ffi.Struct {
+final class rlimit extends ffi.Struct {
   @rlim_t()
   external int rlim_cur;
 
@@ -5002,7 +5182,7 @@ class rlimit extends ffi.Struct {
 
 typedef rlim_t = __uint64_t;
 
-class proc_rlimit_control_wakeupmon extends ffi.Struct {
+final class proc_rlimit_control_wakeupmon extends ffi.Struct {
   @ffi.Uint32()
   external int wm_flags;
 
@@ -5013,9 +5193,9 @@ class proc_rlimit_control_wakeupmon extends ffi.Struct {
 typedef id_t = __darwin_id_t;
 typedef __darwin_id_t = __uint32_t;
 
-class wait extends ffi.Opaque {}
+final class wait extends ffi.Opaque {}
 
-class div_t extends ffi.Struct {
+final class div_t extends ffi.Struct {
   @ffi.Int()
   external int quot;
 
@@ -5023,7 +5203,7 @@ class div_t extends ffi.Struct {
   external int rem;
 }
 
-class ldiv_t extends ffi.Struct {
+final class ldiv_t extends ffi.Struct {
   @ffi.Long()
   external int quot;
 
@@ -5031,7 +5211,7 @@ class ldiv_t extends ffi.Struct {
   external int rem;
 }
 
-class lldiv_t extends ffi.Struct {
+final class lldiv_t extends ffi.Struct {
   @ffi.LongLong()
   external int quot;
 
@@ -5042,7 +5222,7 @@ class lldiv_t extends ffi.Struct {
 typedef malloc_type_id_t = ffi.UnsignedLongLong;
 typedef Dartmalloc_type_id_t = int;
 
-class _malloc_zone_t extends ffi.Opaque {}
+final class _malloc_zone_t extends ffi.Opaque {}
 
 typedef malloc_zone_t = _malloc_zone_t;
 typedef dev_t = __darwin_dev_t;
@@ -5058,25 +5238,56 @@ typedef ssize_t = __darwin_ssize_t;
 typedef __darwin_ssize_t = ffi.Long;
 typedef Dart__darwin_ssize_t = int;
 
-abstract class PLGNStatusCode {
-  static const int PLGNSTATUSCODE_ERROR = 0;
-  static const int PLGNSTATUSCODE_NIL_POINTER = 1;
+enum PLGNStatusCode {
+  PLGNSTATUSCODE_ERROR(0),
+  PLGNSTATUSCODE_NIL_POINTER(1),
+  PLGNSTATUSCODE_USER_CREDENTIAL_STATUS_EXTRACTION_ERROR(2),
+  PLGNSTATUSCODE_USER_CREDENTIAL_STATUS_RESOLVE_ERROR(3),
+  PLGNSTATUSCODE_USER_CREDENTIAL_STATUS_MT_BUILD_ERROR(4),
+  PLGNSTATUSCODE_USER_CREDENTIAL_STATUS_MT_STATE_ERROR(5),
+  PLGNSTATUSCODE_USER_CREDENTIAL_STATUS_REVOKED_ERROR(6),
+  PLGNSTATUSCODE_ISSUER_CREDENTIAL_STATUS_EXTRACTION_ERROR(7),
+  PLGNSTATUSCODE_ISSUER_CREDENTIAL_STATUS_RESOLVE_ERROR(8),
+  PLGNSTATUSCODE_ISSUER_CREDENTIAL_STATUS_MT_BUILD_ERROR(9),
+  PLGNSTATUSCODE_ISSUER_CREDENTIAL_STATUS_MT_STATE_ERROR(10),
+  PLGNSTATUSCODE_ISSUER_CREDENTIAL_STATUS_REVOKED_ERROR(11);
+
+  final int value;
+  const PLGNStatusCode(this.value);
+
+  static PLGNStatusCode fromValue(int value) => switch (value) {
+        0 => PLGNSTATUSCODE_ERROR,
+        1 => PLGNSTATUSCODE_NIL_POINTER,
+        2 => PLGNSTATUSCODE_USER_CREDENTIAL_STATUS_EXTRACTION_ERROR,
+        3 => PLGNSTATUSCODE_USER_CREDENTIAL_STATUS_RESOLVE_ERROR,
+        4 => PLGNSTATUSCODE_USER_CREDENTIAL_STATUS_MT_BUILD_ERROR,
+        5 => PLGNSTATUSCODE_USER_CREDENTIAL_STATUS_MT_STATE_ERROR,
+        6 => PLGNSTATUSCODE_USER_CREDENTIAL_STATUS_REVOKED_ERROR,
+        7 => PLGNSTATUSCODE_ISSUER_CREDENTIAL_STATUS_EXTRACTION_ERROR,
+        8 => PLGNSTATUSCODE_ISSUER_CREDENTIAL_STATUS_RESOLVE_ERROR,
+        9 => PLGNSTATUSCODE_ISSUER_CREDENTIAL_STATUS_MT_BUILD_ERROR,
+        10 => PLGNSTATUSCODE_ISSUER_CREDENTIAL_STATUS_MT_STATE_ERROR,
+        11 => PLGNSTATUSCODE_ISSUER_CREDENTIAL_STATUS_REVOKED_ERROR,
+        _ => throw ArgumentError("Unknown value for PLGNStatusCode: $value"),
+      };
 }
 
-class _PLGNStatus extends ffi.Struct {
-  @ffi.Int32()
-  external int status;
+final class _PLGNStatus extends ffi.Struct {
+  @ffi.UnsignedInt()
+  external int statusAsInt;
+
+  PLGNStatusCode get status => PLGNStatusCode.fromValue(statusAsInt);
 
   external ffi.Pointer<ffi.Char> error_msg;
 }
 
-class GoInterface extends ffi.Struct {
+final class GoInterface extends ffi.Struct {
   external ffi.Pointer<ffi.Void> t;
 
   external ffi.Pointer<ffi.Void> v;
 }
 
-class GoSlice extends ffi.Struct {
+final class GoSlice extends ffi.Struct {
   external ffi.Pointer<ffi.Void> data;
 
   @GoInt()
@@ -5305,6 +5516,8 @@ const int __MAC_14_5 = 140500;
 
 const int __MAC_15_0 = 150000;
 
+const int __MAC_15_1 = 150100;
+
 const int __IPHONE_2_0 = 20000;
 
 const int __IPHONE_2_1 = 20100;
@@ -5465,6 +5678,8 @@ const int __IPHONE_17_5 = 170500;
 
 const int __IPHONE_18_0 = 180000;
 
+const int __IPHONE_18_1 = 180100;
+
 const int __WATCHOS_1_0 = 10000;
 
 const int __WATCHOS_2_0 = 20000;
@@ -5560,6 +5775,8 @@ const int __WATCHOS_10_4 = 100400;
 const int __WATCHOS_10_5 = 100500;
 
 const int __WATCHOS_11_0 = 110000;
+
+const int __WATCHOS_11_1 = 110100;
 
 const int __TVOS_9_0 = 90000;
 
@@ -5659,6 +5876,8 @@ const int __TVOS_17_5 = 170500;
 
 const int __TVOS_18_0 = 180000;
 
+const int __TVOS_18_1 = 180100;
+
 const int __BRIDGEOS_2_0 = 20000;
 
 const int __BRIDGEOS_3_0 = 30000;
@@ -5713,6 +5932,8 @@ const int __BRIDGEOS_8_5 = 80500;
 
 const int __BRIDGEOS_9_0 = 90000;
 
+const int __BRIDGEOS_9_1 = 90100;
+
 const int __DRIVERKIT_19_0 = 190000;
 
 const int __DRIVERKIT_20_0 = 200000;
@@ -5741,6 +5962,8 @@ const int __DRIVERKIT_23_5 = 230500;
 
 const int __DRIVERKIT_24_0 = 240000;
 
+const int __DRIVERKIT_24_1 = 240100;
+
 const int __VISIONOS_1_0 = 10000;
 
 const int __VISIONOS_1_1 = 10100;
@@ -5748,6 +5971,8 @@ const int __VISIONOS_1_1 = 10100;
 const int __VISIONOS_1_2 = 10200;
 
 const int __VISIONOS_2_0 = 20000;
+
+const int __VISIONOS_2_1 = 20100;
 
 const int MAC_OS_X_VERSION_10_0 = 1000;
 
@@ -5873,9 +6098,11 @@ const int MAC_OS_VERSION_14_5 = 140500;
 
 const int MAC_OS_VERSION_15_0 = 150000;
 
-const int __MAC_OS_X_VERSION_MIN_REQUIRED = 150000;
+const int MAC_OS_VERSION_15_1 = 150100;
 
-const int __MAC_OS_X_VERSION_MAX_ALLOWED = 150000;
+const int __MAC_OS_X_VERSION_MIN_REQUIRED = 140000;
+
+const int __MAC_OS_X_VERSION_MAX_ALLOWED = 150100;
 
 const int __ENABLE_LEGACY_MAC_AVAILABILITY = 1;
 
