@@ -1,11 +1,5 @@
-import 'dart:convert';
-import 'dart:math';
-
 import 'package:polygonid_flutter_sdk/assets/get_issuer_id_interface.g.dart';
 import 'package:polygonid_flutter_sdk/assets/onchain_non_merkelized_issuer_base.g.dart';
-import 'package:polygonid_flutter_sdk/common/domain/domain_constants.dart';
-import 'package:polygonid_flutter_sdk/common/domain/entities/chain_config_entity.dart';
-import 'package:polygonid_flutter_sdk/common/domain/entities/env_entity.dart';
 import 'package:polygonid_flutter_sdk/common/domain/error_exception.dart';
 import 'package:polygonid_flutter_sdk/common/domain/use_cases/get_env_use_case.dart';
 import 'package:polygonid_flutter_sdk/common/domain/use_cases/get_selected_chain_use_case.dart';
@@ -13,10 +7,8 @@ import 'package:polygonid_flutter_sdk/common/infrastructure/stacktrace_stream_ma
 import 'package:polygonid_flutter_sdk/credential/domain/use_cases/cache_credential_use_case.dart';
 import 'package:polygonid_flutter_sdk/iden3comm/domain/entities/credential/request/base.dart';
 import 'package:polygonid_flutter_sdk/iden3comm/domain/entities/credential/request/onchain_offer_iden3_message_entity.dart';
-import 'package:polygonid_flutter_sdk/iden3comm/domain/entities/interaction/interaction_entity.dart';
 import 'package:polygonid_flutter_sdk/iden3comm/domain/exceptions/iden3comm_exceptions.dart';
 import 'package:polygonid_flutter_sdk/iden3comm/domain/repositories/iden3comm_credential_repository.dart';
-import 'package:polygonid_flutter_sdk/iden3comm/domain/repositories/interaction_repository.dart';
 import 'package:polygonid_flutter_sdk/iden3comm/domain/use_cases/check_profile_and_did_current_env.dart';
 import 'package:polygonid_flutter_sdk/iden3comm/domain/use_cases/fetch_onchain_claim_use_case.dart';
 import 'package:polygonid_flutter_sdk/identity/data/data_sources/local_contract_files_data_source.dart';
@@ -32,7 +24,6 @@ import 'package:polygonid_flutter_sdk/iden3comm/domain/entities/credential/reque
 import 'package:polygonid_flutter_sdk/iden3comm/domain/use_cases/get_auth_token_use_case.dart';
 import 'package:polygonid_flutter_sdk/iden3comm/domain/use_cases/get_fetch_requests_use_case.dart';
 import 'package:polygonid_flutter_sdk/identity/domain/use_cases/get_did_use_case.dart';
-import 'package:polygonid_flutter_sdk/proof/data/dtos/atomic_query_inputs_config_param.dart';
 import 'package:polygonid_flutter_sdk/sdk/di/injector.dart';
 import 'package:web3dart/crypto.dart';
 import 'package:web3dart/web3dart.dart';
@@ -67,7 +58,6 @@ class FetchAndSaveClaimsUseCase
   final CacheCredentialUseCase _cacheCredentialUseCase;
   final LocalContractFilesDataSource _localContractFilesDataSource;
   final IdentityRepository _identityRepository;
-  final InteractionRepository _interactionRepository;
 
   final StacktraceManager _stacktraceManager;
 
@@ -85,7 +75,6 @@ class FetchAndSaveClaimsUseCase
     this._cacheCredentialUseCase,
     this._localContractFilesDataSource,
     this._identityRepository,
-    this._interactionRepository,
     this._stacktraceManager,
   );
 
@@ -146,12 +135,7 @@ class FetchAndSaveClaimsUseCase
       _stacktraceManager.addTrace(
           "[FetchAndSaveClaimsUseCase] All claims have been saved: claimsLength ${claims.length}");
 
-      final config = ConfigParam(
-        ipfsNodeURL: env.ipfsUrl,
-        chainConfigs: env.chainConfigs,
-        didMethods: env.didMethods,
-      );
-
+      final config = env.config;
       for (final claim in claims) {
         // cache claim
         try {
@@ -263,7 +247,7 @@ class FetchAndSaveClaimsUseCase
     final issuerIdInt = await getIssuerId.getId();
     final issuerDid = (await _identityRepository.describeId(
       id: issuerIdInt,
-      config: ConfigParam.fromEnv(env),
+      config: env.config,
     ))
         .did;
 
@@ -309,35 +293,5 @@ class FetchAndSaveClaimsUseCase
     }
 
     return claims;
-  }
-
-  Future<BigInt> _tryFindLastInteractionWithPrivateProfile(
-    String genesisDid,
-    String privateKey,
-    String issuerDid,
-  ) async {
-    // Public DID nonce
-    BigInt profileNonce = BigInt.zero;
-
-    var interactions = await _interactionRepository.getInteractions(
-      encryptionKey: privateKey,
-      genesisDid: genesisDid,
-    );
-
-    interactions = interactions
-        .where((interaction) => interaction.from == issuerDid)
-        .toList();
-    interactions.sort((a, b) => max(a.timestamp, b.timestamp));
-
-    if (interactions.isEmpty) {
-      return GENESIS_PROFILE_NONCE;
-    }
-
-    final latestInteraction = interactions.last;
-    if (latestInteraction is InteractionEntity) {
-      profileNonce = latestInteraction.profileNonce;
-    }
-
-    return profileNonce;
   }
 }
