@@ -4,10 +4,7 @@ import 'package:polygonid_flutter_sdk/common/domain/use_case.dart';
 import 'package:polygonid_flutter_sdk/common/infrastructure/stacktrace_stream_manager.dart';
 import 'package:polygonid_flutter_sdk/common/utils/credential_sort_order.dart';
 import 'package:polygonid_flutter_sdk/credential/domain/entities/claim_entity.dart';
-import 'package:polygonid_flutter_sdk/credential/domain/use_cases/get_claim_revocation_nonce_use_case.dart';
-import 'package:polygonid_flutter_sdk/credential/domain/use_cases/get_claim_revocation_status_use_case.dart';
 import 'package:polygonid_flutter_sdk/credential/domain/use_cases/get_claims_use_case.dart';
-import 'package:polygonid_flutter_sdk/credential/domain/use_cases/update_claim_use_case.dart';
 import 'package:polygonid_flutter_sdk/iden3comm/domain/entities/common/iden3_message_entity.dart';
 import 'package:polygonid_flutter_sdk/iden3comm/domain/entities/common/request/proof_request_entity.dart';
 import 'package:polygonid_flutter_sdk/iden3comm/domain/entities/common/request/proof_scope_query_request.dart';
@@ -15,7 +12,6 @@ import 'package:polygonid_flutter_sdk/iden3comm/domain/exceptions/iden3comm_exce
 import 'package:polygonid_flutter_sdk/iden3comm/domain/repositories/iden3comm_credential_repository.dart';
 import 'package:polygonid_flutter_sdk/iden3comm/domain/use_cases/get_proof_requests_use_case.dart';
 import 'package:polygonid_flutter_sdk/identity/data/dtos/circuit_type.dart';
-import 'package:polygonid_flutter_sdk/proof/data/mappers/circuit_type_mapper.dart';
 import 'package:polygonid_flutter_sdk/proof/domain/exceptions/proof_generation_exceptions.dart';
 import 'package:polygonid_flutter_sdk/proof/domain/use_cases/is_proof_circuit_supported_use_case.dart';
 
@@ -42,23 +38,15 @@ class GetIden3commClaimsUseCase
     extends FutureUseCase<GetIden3commClaimsParam, List<ClaimEntity?>> {
   final Iden3commCredentialRepository _iden3commCredentialRepository;
   final GetClaimsUseCase _getClaimsUseCase;
-  final GetClaimRevocationStatusUseCase _getClaimRevocationStatusUseCase;
-  final GetClaimRevocationNonceUseCase _getClaimRevocationNonceUseCase;
-  final UpdateClaimUseCase _updateClaimUseCase;
   final IsProofCircuitSupportedUseCase _isProofCircuitSupported;
   final GetProofRequestsUseCase _getProofRequestsUseCase;
-  final CircuitTypeMapper _circuitTypeMapper;
   final StacktraceManager _stacktraceManager;
 
   GetIden3commClaimsUseCase(
     this._iden3commCredentialRepository,
     this._getClaimsUseCase,
-    this._getClaimRevocationStatusUseCase,
-    this._getClaimRevocationNonceUseCase,
-    this._updateClaimUseCase,
     this._isProofCircuitSupported,
     this._getProofRequestsUseCase,
-    this._circuitTypeMapper,
     this._stacktraceManager,
   );
 
@@ -177,7 +165,7 @@ class GetIden3commClaimsUseCase
                   "V3 circuit beta version mismatch $circuitId is not supported, current is ${CircuitType.currentCircuitBetaPostfix}");
         }
 
-        CircuitType circuitType = _circuitTypeMapper.mapTo(circuitId);
+        CircuitType circuitType = CircuitType.fromString(circuitId);
 
         switch (circuitType) {
           case CircuitType.mtp:
