@@ -1,3 +1,4 @@
+import 'package:flutter/foundation.dart';
 import 'package:injectable/injectable.dart';
 import 'package:polygonid_flutter_sdk/common/data/data_sources/secure_identity_storage_data_source.dart';
 import 'package:sembast/sembast.dart';
@@ -17,18 +18,28 @@ class IdentitySMTStoreRefWrapper {
   IdentitySMTStoreRefWrapper(@Named(identityStateStoreName) this._store);
 
   Future<Map<String, Object?>?> get(
-      DatabaseClient database, String storeName, String key) {
+    DatabaseClient database,
+    String storeName,
+    String key,
+  ) {
     return _store[storeName]!.record(key).get(database);
   }
 
-  Future<Map<String, Object?>> put(DatabaseClient database, String storeName,
-      String key, Map<String, Object?> value,
-      {bool? merge}) {
+  Future<Map<String, Object?>> put(
+    DatabaseClient database,
+    String storeName,
+    String key,
+    Map<String, Object?> value, {
+    bool? merge,
+  }) {
     return _store[storeName]!.record(key).put(database, value, merge: merge);
   }
 
   Future<String?> remove(
-      DatabaseClient database, String storeName, String did) {
+    DatabaseClient database,
+    String storeName,
+    String did,
+  ) {
     return _store[storeName]!.record(did).delete(database);
   }
 
@@ -49,13 +60,13 @@ class StorageSMTDataSource extends SecureIdentityStorageDataSource {
     required String storeName,
     required String did,
     required String encryptionKey,
-  }) {
-    if ((BigInt.parse(key.toString())) == BigInt.zero) {
-      return Future.value(NodeEntity(
+  }) async {
+    if (key.toBigInt() == BigInt.zero) {
+      return NodeEntity(
         children: const [],
         hash: HashEntity.fromBigInt(BigInt.zero),
         type: NodeType.empty,
-      ));
+      );
     }
     return getDatabase(did: did, encryptionKey: encryptionKey).then(
         (database) => database
@@ -65,7 +76,7 @@ class StorageSMTDataSource extends SecureIdentityStorageDataSource {
             .whenComplete(() => database.close()));
   }
 
-  // For UT purpose
+  @visibleForTesting
   Future<Map<String, Object?>?> getTransact({
     required String storeName,
     required DatabaseClient transaction,
@@ -81,8 +92,8 @@ class StorageSMTDataSource extends SecureIdentityStorageDataSource {
     required String did,
     required String encryptionKey,
   }) {
-    return getDatabase(did: did, encryptionKey: encryptionKey).then(
-        (database) => database
+    return getDatabase(did: did, encryptionKey: encryptionKey).then((database) =>
+        database
             .transaction((transaction) => putTransact(
                 storeName: storeName,
                 transaction: transaction,
@@ -91,7 +102,7 @@ class StorageSMTDataSource extends SecureIdentityStorageDataSource {
             .whenComplete(() => database.close()));
   }
 
-  // For UT purpose
+  @visibleForTesting
   Future<void> putTransact({
     required String storeName,
     required DatabaseClient transaction,
@@ -99,7 +110,11 @@ class StorageSMTDataSource extends SecureIdentityStorageDataSource {
     required NodeEntity node,
   }) async {
     await _storeRefWrapper.put(
-        transaction, storeName, key.toString(), node.toJson());
+      transaction,
+      storeName,
+      key.toString(),
+      node.toJson(),
+    );
   }
 
   Future<HashEntity> getRoot({
@@ -107,16 +122,18 @@ class StorageSMTDataSource extends SecureIdentityStorageDataSource {
     required String did,
     required String encryptionKey,
   }) {
-    return getDatabase(did: did, encryptionKey: encryptionKey).then(
-        (database) => database
+    return getDatabase(did: did, encryptionKey: encryptionKey).then((database) =>
+        database
             .transaction((transaction) =>
                 getRootTransact(transaction: transaction, storeName: storeName))
             .whenComplete(() => database.close()));
   }
 
-  // For UT purpose
-  Future<HashEntity> getRootTransact(
-      {required DatabaseClient transaction, required String storeName}) async {
+  @visibleForTesting
+  Future<HashEntity> getRootTransact({
+    required DatabaseClient transaction,
+    required String storeName,
+  }) async {
     return _storeRefWrapper
         .get(transaction, storeName, "root")
         .then((storedValue) {
@@ -136,14 +153,14 @@ class StorageSMTDataSource extends SecureIdentityStorageDataSource {
     required String did,
     required String encryptionKey,
   }) {
-    return getDatabase(did: did, encryptionKey: encryptionKey).then(
-        (database) => database
+    return getDatabase(did: did, encryptionKey: encryptionKey).then((database) =>
+        database
             .transaction((transaction) => setRootTransact(
                 transaction: transaction, storeName: storeName, root: root))
             .whenComplete(() => database.close()));
   }
 
-  // For UT purpose
+  @visibleForTesting
   Future<void> setRootTransact({
     required DatabaseClient transaction,
     required String storeName,
@@ -156,17 +173,17 @@ class StorageSMTDataSource extends SecureIdentityStorageDataSource {
   Future<int> getMaxLevels({
     required String storeName,
     required String did,
-    required String encryptionKey,
+    required String privateKey,
   }) {
-    return getDatabase(did: did, encryptionKey: encryptionKey).then(
-        (database) => database
+    return getDatabase(did: did, encryptionKey: encryptionKey).then((database) =>
+        database
             .transaction((transaction) => getMaxLevelsTransact(
                 transaction: transaction, storeName: storeName))
             .then((snapshot) => snapshot!["maxLevels"] as int)
             .whenComplete(() => database.close()));
   }
 
-  // For UT purpose
+  @visibleForTesting
   Future<Map<String, Object?>?> getMaxLevelsTransact({
     required DatabaseClient transaction,
     required String storeName,
@@ -180,8 +197,8 @@ class StorageSMTDataSource extends SecureIdentityStorageDataSource {
     required String did,
     required String encryptionKey,
   }) {
-    return getDatabase(did: did, encryptionKey: encryptionKey).then(
-        (database) => database
+    return getDatabase(did: did, encryptionKey: encryptionKey).then((database) =>
+        database
             .transaction((transaction) => setMaxLevelsTransact(
                 transaction: transaction,
                 storeName: storeName,
@@ -189,11 +206,12 @@ class StorageSMTDataSource extends SecureIdentityStorageDataSource {
             .whenComplete(() => database.close()));
   }
 
-  // For UT purpose
-  Future<void> setMaxLevelsTransact(
-      {required DatabaseClient transaction,
-      required String storeName,
-      required int maxLevels}) async {
+  @visibleForTesting
+  Future<void> setMaxLevelsTransact({
+    required DatabaseClient transaction,
+    required String storeName,
+    required int maxLevels,
+  }) async {
     await _storeRefWrapper
         .put(transaction, storeName, "maxLevels", {"maxLevels": maxLevels});
   }
@@ -203,14 +221,14 @@ class StorageSMTDataSource extends SecureIdentityStorageDataSource {
     required String did,
     required String encryptionKey,
   }) {
-    return getDatabase(did: did, encryptionKey: encryptionKey).then(
-        (database) => database
+    return getDatabase(did: did, encryptionKey: encryptionKey).then((database) =>
+        database
             .transaction((transaction) => removeSMTTransact(
                 transaction: transaction, storeName: storeName))
             .whenComplete(() => database.close()));
   }
 
-  // For UT purpose
+  @visibleForTesting
   Future<void> removeSMTTransact({
     required DatabaseClient transaction,
     required String storeName,
