@@ -1,5 +1,11 @@
+import 'package:polygonid_flutter_sdk/constants.dart';
+
 enum TreeType {
-  claims,
-  revocation,
-  roots,
+  claims(claimsTreeStoreName),
+  revocation(revocationTreeStoreName),
+  roots(rootsTreeStoreName);
+
+  final String storeName;
+
+  const TreeType(this.storeName);
 }
