@@ -1,45 +1,49 @@
 import 'dart:convert';
-import 'dart:io';
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
-import 'package:injectable/injectable.dart';
-import 'package:polygonid_flutter_sdk/proof/data/dtos/prove_param.dart';
-import 'package:polygonid_flutter_sdk/proof/libs/prover/prover.dart';
+import 'package:flutter_rapidsnark/flutter_rapidsnark.dart';
 
-const _methodChannel = MethodChannel('polygonid_flutter_sdk');
+class ProveParam {
+  final String zKeyPath;
+  final Uint8List wtns;
+  final RootIsolateToken rootToken;
 
-@injectable
-class ProverLibWrapper {
-  Future<Map<String, dynamic>?> prover(
-      String circuitId, String zKeyPath, Uint8List wtnsBytes) {
-    return compute(_computeProve, ProveParam(circuitId, zKeyPath, wtnsBytes));
-  }
-}
-
-///
-Future<Map<String, dynamic>?> _computeProve(ProveParam param) async {
-  ProverLib proverLib = ProverLib();
-  return proverLib.proveZkeyFilePath(
-    param.circuitId,
-    param.zKeyPath,
-    param.wtns,
+  ProveParam(
+    this.zKeyPath,
+    this.wtns,
+    this.rootToken,
   );
 }
 
 class ProverLibDataSource {
-  final ProverLibWrapper _proverLibWrapper;
-
-  ProverLibDataSource(this._proverLibWrapper);
+  ProverLibDataSource();
 
   ///
   Future<Map<String, dynamic>?> prove(
-    String circuitId,
     String zKeyPath,
     Uint8List wtnsBytes,
   ) async {
-    final result =
-        await _proverLibWrapper.prover(circuitId, zKeyPath, wtnsBytes);
-    return result;
+    final rootToken = RootIsolateToken.instance!;
+
+    return compute(
+      _computeProof,
+      ProveParam(zKeyPath, wtnsBytes, rootToken),
+    );
   }
+}
+
+///
+Future<Map<String, dynamic>?> _computeProof(ProveParam param) async {
+  BackgroundIsolateBinaryMessenger.ensureInitialized(param.rootToken);
+
+  final result = await Rapidsnark().groth16Prove(
+    zkeyPath: param.zKeyPath,
+    witness: param.wtns,
+  );
+
+  return {
+    'proof': jsonDecode(result.proof),
+    'pub_signals': jsonDecode(result.publicSignals),
+  };
 }
