@@ -23,20 +23,12 @@ import 'package:polygonid_flutter_sdk/common/data/data_sources/mappers/filter_ma
     as _i325;
 import 'package:polygonid_flutter_sdk/common/data/data_sources/mappers/filters_mapper.dart'
     as _i461;
-import 'package:polygonid_flutter_sdk/common/data/data_sources/package_info_datasource.dart'
-    as _i355;
 import 'package:polygonid_flutter_sdk/common/data/data_sources/storage_key_value_data_source.dart'
     as _i525;
-import 'package:polygonid_flutter_sdk/common/data/repositories/config_repository_impl.dart'
-    as _i1049;
-import 'package:polygonid_flutter_sdk/common/data/repositories/package_info_repository_impl.dart'
-    as _i47;
 import 'package:polygonid_flutter_sdk/common/domain/domain_logger.dart'
     as _i1041;
 import 'package:polygonid_flutter_sdk/common/domain/repositories/config_repository.dart'
     as _i415;
-import 'package:polygonid_flutter_sdk/common/domain/repositories/package_info_repository.dart'
-    as _i141;
 import 'package:polygonid_flutter_sdk/common/domain/use_cases/get_env_use_case.dart'
     as _i626;
 import 'package:polygonid_flutter_sdk/common/domain/use_cases/get_package_name_use_case.dart'
@@ -49,7 +41,6 @@ import 'package:polygonid_flutter_sdk/common/domain/use_cases/set_selected_chain
     as _i438;
 import 'package:polygonid_flutter_sdk/common/infrastructure/stacktrace_stream_manager.dart'
     as _i267;
-import 'package:polygonid_flutter_sdk/common/kms/index.dart' as _i710;
 import 'package:polygonid_flutter_sdk/common/libs/polygonidcore/pidcore_base.dart'
     as _i393;
 import 'package:polygonid_flutter_sdk/credential/data/credential_repository_impl.dart'
@@ -72,10 +63,6 @@ import 'package:polygonid_flutter_sdk/credential/data/mappers/claim_state_mapper
     as _i497;
 import 'package:polygonid_flutter_sdk/credential/data/mappers/display_type_mapper.dart'
     as _i590;
-import 'package:polygonid_flutter_sdk/credential/data/mappers/id_filter_mapper.dart'
-    as _i161;
-import 'package:polygonid_flutter_sdk/credential/data/mappers/revocation_status_mapper.dart'
-    as _i162;
 import 'package:polygonid_flutter_sdk/credential/domain/repositories/credential_repository.dart'
     as _i309;
 import 'package:polygonid_flutter_sdk/credential/domain/use_cases/add_did_profile_info_use_case.dart'
@@ -118,8 +105,6 @@ import 'package:polygonid_flutter_sdk/credential/libs/polygonidcore/pidcore_cred
     as _i328;
 import 'package:polygonid_flutter_sdk/iden3comm/data/data_sources/iden3_message_data_source.dart'
     as _i296;
-import 'package:polygonid_flutter_sdk/iden3comm/data/data_sources/lib_pidcore_iden3comm_data_source.dart'
-    as _i882;
 import 'package:polygonid_flutter_sdk/iden3comm/data/data_sources/remote_iden3comm_data_source.dart'
     as _i409;
 import 'package:polygonid_flutter_sdk/iden3comm/data/data_sources/secure_storage_did_profile_info_data_source.dart'
@@ -128,16 +113,6 @@ import 'package:polygonid_flutter_sdk/iden3comm/data/data_sources/secure_storage
     as _i425;
 import 'package:polygonid_flutter_sdk/iden3comm/data/data_sources/storage_interaction_data_source.dart'
     as _i57;
-import 'package:polygonid_flutter_sdk/iden3comm/data/mappers/auth_proof_mapper.dart'
-    as _i1050;
-import 'package:polygonid_flutter_sdk/iden3comm/data/mappers/auth_response_mapper.dart'
-    as _i22;
-import 'package:polygonid_flutter_sdk/iden3comm/data/mappers/did_profile_info_interacted_did_filter_mapper.dart'
-    as _i435;
-import 'package:polygonid_flutter_sdk/iden3comm/data/mappers/iden3_message_type_mapper.dart'
-    as _i284;
-import 'package:polygonid_flutter_sdk/iden3comm/data/mappers/interaction_id_filter_mapper.dart'
-    as _i609;
 import 'package:polygonid_flutter_sdk/iden3comm/data/mappers/interaction_mapper.dart'
     as _i1026;
 import 'package:polygonid_flutter_sdk/iden3comm/data/mappers/jwz_mapper.dart'
@@ -212,10 +187,6 @@ import 'package:polygonid_flutter_sdk/iden3comm/domain/use_cases/interaction/rem
     as _i975;
 import 'package:polygonid_flutter_sdk/iden3comm/domain/use_cases/interaction/update_interaction_use_case.dart'
     as _i989;
-import 'package:polygonid_flutter_sdk/iden3comm/domain/use_cases/listen_and_store_notification_use_case.dart'
-    as _i191;
-import 'package:polygonid_flutter_sdk/iden3comm/libs/polygonidcore/pidcore_iden3comm.dart'
-    as _i360;
 import 'package:polygonid_flutter_sdk/identity/data/data_sources/db_destination_path_data_source.dart'
     as _i938;
 import 'package:polygonid_flutter_sdk/identity/data/data_sources/encryption_db_data_source.dart'
@@ -238,8 +209,6 @@ import 'package:polygonid_flutter_sdk/identity/data/data_sources/storage_smt_dat
     as _i42;
 import 'package:polygonid_flutter_sdk/identity/data/data_sources/wallet_data_source.dart'
     as _i383;
-import 'package:polygonid_flutter_sdk/identity/data/mappers/hex_mapper.dart'
-    as _i107;
 import 'package:polygonid_flutter_sdk/identity/data/mappers/node_type_entity_mapper.dart'
     as _i307;
 import 'package:polygonid_flutter_sdk/identity/data/mappers/private_key/private_key_hex_mapper.dart'
@@ -250,16 +219,8 @@ import 'package:polygonid_flutter_sdk/identity/data/mappers/private_key/private_
     as _i168;
 import 'package:polygonid_flutter_sdk/identity/data/mappers/q_mapper.dart'
     as _i599;
-import 'package:polygonid_flutter_sdk/identity/data/mappers/rhs_node_mapper.dart'
-    as _i340;
-import 'package:polygonid_flutter_sdk/identity/data/mappers/rhs_node_type_mapper.dart'
-    as _i1049;
 import 'package:polygonid_flutter_sdk/identity/data/mappers/state_identifier_mapper.dart'
     as _i720;
-import 'package:polygonid_flutter_sdk/identity/data/mappers/tree_state_mapper.dart'
-    as _i68;
-import 'package:polygonid_flutter_sdk/identity/data/mappers/tree_type_mapper.dart'
-    as _i443;
 import 'package:polygonid_flutter_sdk/identity/data/repositories/identity_repository_impl.dart'
     as _i393;
 import 'package:polygonid_flutter_sdk/identity/data/repositories/smt_repository_impl.dart'
@@ -343,12 +304,6 @@ import 'package:polygonid_flutter_sdk/proof/data/data_sources/prover_lib_data_so
     as _i502;
 import 'package:polygonid_flutter_sdk/proof/data/data_sources/witness_data_source.dart'
     as _i1039;
-import 'package:polygonid_flutter_sdk/proof/data/mappers/circuit_type_mapper.dart'
-    as _i512;
-import 'package:polygonid_flutter_sdk/proof/data/mappers/gist_mtproof_mapper.dart'
-    as _i178;
-import 'package:polygonid_flutter_sdk/proof/data/mappers/zkproof_mapper.dart'
-    as _i360;
 import 'package:polygonid_flutter_sdk/proof/data/repositories/proof_repository_impl.dart'
     as _i581;
 import 'package:polygonid_flutter_sdk/proof/domain/repositories/proof_repository.dart'
@@ -373,23 +328,6 @@ import 'package:polygonid_flutter_sdk/proof/infrastructure/proof_generation_stre
     as _i920;
 import 'package:polygonid_flutter_sdk/proof/libs/polygonidcore/pidcore_proof.dart'
     as _i961;
-import 'package:polygonid_flutter_sdk/proof/libs/prover/prover.dart' as _i506;
-import 'package:polygonid_flutter_sdk/proof/libs/witnesscalc/auth_v2/witness_auth.dart'
-    as _i318;
-import 'package:polygonid_flutter_sdk/proof/libs/witnesscalc/linked_multi_query_10/witness_linked_multi_query_10.dart'
-    as _i896;
-import 'package:polygonid_flutter_sdk/proof/libs/witnesscalc/mtp_v2/witness_mtp.dart'
-    as _i569;
-import 'package:polygonid_flutter_sdk/proof/libs/witnesscalc/mtp_v2_onchain/witness_mtp_onchain.dart'
-    as _i436;
-import 'package:polygonid_flutter_sdk/proof/libs/witnesscalc/sig_v2/witness_sig.dart'
-    as _i184;
-import 'package:polygonid_flutter_sdk/proof/libs/witnesscalc/sig_v2_onchain/witness_sig_onchain.dart'
-    as _i68;
-import 'package:polygonid_flutter_sdk/proof/libs/witnesscalc/v3/witness_v3.dart'
-    as _i695;
-import 'package:polygonid_flutter_sdk/proof/libs/witnesscalc/v3_onchain/witness_v3_onchain.dart'
-    as _i602;
 import 'package:polygonid_flutter_sdk/sdk/credential.dart' as _i501;
 import 'package:polygonid_flutter_sdk/sdk/di/injector.dart' as _i335;
 import 'package:polygonid_flutter_sdk/sdk/error_handling.dart' as _i795;
@@ -417,39 +355,19 @@ extension GetItInjectableX on _i174.GetIt {
     final channelModule = _$ChannelModule();
     final networkModule = _$NetworkModule();
     final filesManagerModule = _$FilesManagerModule();
-    final kMSModule = _$KMSModule();
     final platformModule = _$PlatformModule();
     final databaseModule = _$DatabaseModule();
     final encryptionModule = _$EncryptionModule();
     final repositoriesModule = _$RepositoriesModule();
-    gh.factory<_i318.WitnessAuthV2Lib>(() => _i318.WitnessAuthV2Lib());
-    gh.factory<_i602.WitnessV3OnchainLib>(() => _i602.WitnessV3OnchainLib());
-    gh.factory<_i184.WitnessSigV2Lib>(() => _i184.WitnessSigV2Lib());
-    gh.factory<_i695.WitnessV3Lib>(() => _i695.WitnessV3Lib());
-    gh.factory<_i896.WitnessLinkedMultiQuery10>(
-        () => _i896.WitnessLinkedMultiQuery10());
-    gh.factory<_i68.WitnessSigV2OnchainLib>(
-        () => _i68.WitnessSigV2OnchainLib());
-    gh.factory<_i569.WitnessMTPV2Lib>(() => _i569.WitnessMTPV2Lib());
-    gh.factory<_i436.WitnessMTPV2OnchainLib>(
-        () => _i436.WitnessMTPV2OnchainLib());
-    gh.factory<_i506.ProverLib>(() => _i506.ProverLib());
-    gh.factory<_i502.ProverLibWrapper>(() => _i502.ProverLibWrapper());
-    gh.factory<_i1039.WitnessIsolatesWrapper>(
-        () => _i1039.WitnessIsolatesWrapper());
+    gh.factory<_i502.ProverLibDataSource>(() => _i502.ProverLibDataSource());
+    gh.factory<_i1039.WitnessDataSource>(() => _i1039.WitnessDataSource());
     gh.factory<_i694.GistMTProofDataSource>(
         () => _i694.GistMTProofDataSource());
     gh.factory<_i1021.ProofCircuitDataSource>(
         () => _i1021.ProofCircuitDataSource());
-    gh.factory<_i178.GistMTProofMapper>(() => _i178.GistMTProofMapper());
-    gh.factory<_i512.CircuitTypeMapper>(() => _i512.CircuitTypeMapper());
-    gh.factory<_i360.ZKProofMapper>(() => _i360.ZKProofMapper());
     gh.factory<_i894.ClaimInfoMapper>(() => _i894.ClaimInfoMapper());
     gh.factory<_i497.ClaimStateMapper>(() => _i497.ClaimStateMapper());
-    gh.factory<_i162.RevocationStatusMapper>(
-        () => _i162.RevocationStatusMapper());
     gh.factory<_i590.DisplayTypeMapper>(() => _i590.DisplayTypeMapper());
-    gh.factory<_i161.IdFilterMapper>(() => _i161.IdFilterMapper());
     gh.factory<_i80.BabyjubjubLib>(() => _i80.BabyjubjubLib());
     gh.factory<_i383.WalletLibWrapper>(() => _i383.WalletLibWrapper());
     gh.factory<_i200.EncryptionDbDataSource>(
@@ -458,16 +376,13 @@ extension GetItInjectableX on _i174.GetIt {
     gh.factory<_i22.LocalContractFilesDataSource>(
         () => _i22.LocalContractFilesDataSource());
     gh.factory<_i599.QMapper>(() => _i599.QMapper());
-    gh.factory<_i107.HexMapper>(() => _i107.HexMapper());
-    gh.factory<_i68.TreeStateMapper>(() => _i68.TreeStateMapper());
     gh.factory<_i720.StateIdentifierMapper>(
         () => _i720.StateIdentifierMapper());
     gh.factory<_i180.PrivateKeyHexMapper>(() => _i180.PrivateKeyHexMapper());
     gh.factory<_i307.NodeTypeEntityMapper>(() => _i307.NodeTypeEntityMapper());
-    gh.factory<_i1049.RhsNodeTypeMapper>(() => _i1049.RhsNodeTypeMapper());
-    gh.factory<_i443.TreeTypeMapper>(() => _i443.TreeTypeMapper());
     gh.factory<_i192.CheckProfileValidityUseCase>(
         () => _i192.CheckProfileValidityUseCase());
+    gh.factory<_i78.GetDidUseCase>(() => _i78.GetDidUseCase());
     gh.factory<_i393.PolygonIdCore>(() => _i393.PolygonIdCore());
     gh.factory<_i325.FilterMapper>(() => _i325.FilterMapper());
     gh.factory<_i974.Logger>(() => loggerModule.logger);
@@ -478,18 +393,9 @@ extension GetItInjectableX on _i174.GetIt {
     gh.factoryAsync<_i497.Directory>(
         () => filesManagerModule.applicationDocumentsDirectory);
     gh.factory<_i71.ZipDecoder>(() => filesManagerModule.zipDecoder());
-    gh.factory<_i435.DidProfileInfoInteractedDidFilterMapper>(
-        () => _i435.DidProfileInfoInteractedDidFilterMapper());
-    gh.factory<_i22.AuthResponseMapper>(() => _i22.AuthResponseMapper());
-    gh.factory<_i1050.AuthProofMapper>(() => _i1050.AuthProofMapper());
-    gh.factory<_i609.InteractionIdFilterMapper>(
-        () => _i609.InteractionIdFilterMapper());
     gh.factory<_i1026.InteractionMapper>(() => _i1026.InteractionMapper());
-    gh.factory<_i284.Iden3MessageTypeMapper>(
-        () => _i284.Iden3MessageTypeMapper());
     gh.factory<_i968.GetFetchRequestsUseCase>(
         () => _i968.GetFetchRequestsUseCase());
-    gh.singleton<_i710.KMS>(() => kMSModule.kms);
     gh.lazySingleton<_i920.ProofGenerationStepsStreamManager>(
         () => _i920.ProofGenerationStepsStreamManager());
     gh.lazySingleton<_i267.StacktraceManager>(() => _i267.StacktraceManager());
@@ -497,6 +403,11 @@ extension GetItInjectableX on _i174.GetIt {
     gh.lazySingletonAsync<_i655.PackageInfo>(() => platformModule.packageInfo);
     gh.lazySingleton<_i281.AssetBundle>(() => platformModule.assetBundle);
     gh.lazySingletonAsync<_i310.Database>(() => databaseModule.database());
+    gh.factoryParam<_i310.SembastCodec, String, dynamic>((
+      privateKey,
+      _,
+    ) =>
+        databaseModule.getCodec(privateKey));
     gh.factory<_i310.StoreRef<String, Map<String, Object?>>>(
       () => databaseModule.claimStore,
       instanceName: 'claimStore',
@@ -512,6 +423,17 @@ extension GetItInjectableX on _i174.GetIt {
     gh.factory<Map<String, _i310.StoreRef<String, Map<String, Object?>>>>(
       () => databaseModule.identityStateStore,
       instanceName: 'identityStateStore',
+    );
+    gh.factoryParamAsync<_i310.Database, String?, String?>(
+      (
+        identifier,
+        privateKey,
+      ) =>
+          databaseModule.identityDatabase(
+        identifier,
+        privateKey,
+      ),
+      instanceName: 'polygonIdSdkIdentity',
     );
     gh.factory<_i310.StoreRef<String, Map<String, Object?>>>(
       () => databaseModule.didProfileInfoStore,
@@ -546,8 +468,6 @@ extension GetItInjectableX on _i174.GetIt {
         () => _i839.RemoteIdentityDataSource(gh<_i267.StacktraceManager>()));
     gh.factory<_i795.ErrorHandling>(
         () => _i795.ErrorHandling(gh<_i267.StacktraceManager>()));
-    gh.factory<_i360.PolygonIdCoreIden3comm>(
-        () => _i360.PolygonIdCoreIden3comm(gh<_i267.StacktraceManager>()));
     gh.factory<_i296.Iden3MessageDataSource>(
         () => _i296.Iden3MessageDataSource(gh<_i267.StacktraceManager>()));
     gh.factory<_i96.ProofRequestFiltersMapper>(
@@ -577,11 +497,6 @@ extension GetItInjectableX on _i174.GetIt {
     gh.factory<_i738.ClaimStoreRefWrapper>(() => _i738.ClaimStoreRefWrapper(
         gh<_i310.StoreRef<String, Map<String, Object?>>>(
             instanceName: 'claimStore')));
-    gh.factoryParam<_i310.SembastCodec, String, dynamic>((
-      encryptionKey,
-      _,
-    ) =>
-        databaseModule.getCodec(encryptionKey));
     gh.factory<_i738.StorageClaimDataSource>(
         () => _i738.StorageClaimDataSource(gh<_i738.ClaimStoreRefWrapper>()));
     gh.factory<_i310.StoreRef<String, Map<String, Object?>>>(
@@ -601,8 +516,6 @@ extension GetItInjectableX on _i174.GetIt {
     gh.factory<_i758.LibPolygonIdCoreCredentialDataSource>(() =>
         _i758.LibPolygonIdCoreCredentialDataSource(
             gh<_i328.PolygonIdCoreCredential>()));
-    gh.factory<_i1039.WitnessDataSource>(
-        () => _i1039.WitnessDataSource(gh<_i1039.WitnessIsolatesWrapper>()));
     gh.factory<_i361.SecureDidProfileInfoStoreRefWrapper>(() =>
         _i361.SecureDidProfileInfoStoreRefWrapper(
             gh<_i310.StoreRef<String, Map<String, Object?>>>(
@@ -619,21 +532,8 @@ extension GetItInjectableX on _i174.GetIt {
     gh.factory<_i425.SecureStorageInteractionDataSource>(() =>
         _i425.SecureStorageInteractionDataSource(
             gh<_i425.SecureInteractionStoreRefWrapper>()));
-    gh.factory<_i502.ProverLibDataSource>(
-        () => _i502.ProverLibDataSource(gh<_i502.ProverLibWrapper>()));
     gh.factory<_i383.WalletDataSource>(
         () => _i383.WalletDataSource(gh<_i383.WalletLibWrapper>()));
-    gh.factoryParamAsync<_i310.Database, String?, String?>(
-      (
-        identifier,
-        encryptionKey,
-      ) =>
-          databaseModule.identityDatabase(
-        identifier,
-        encryptionKey,
-      ),
-      instanceName: 'polygonIdSdkIdentity',
-    );
     gh.factory<_i910.Iden3commCredentialRepositoryImpl>(
         () => _i910.Iden3commCredentialRepositoryImpl(
               gh<_i409.RemoteIden3commDataSource>(),
@@ -650,42 +550,38 @@ extension GetItInjectableX on _i174.GetIt {
               gh<_i189.PolygonIdSdk>(),
               gh<_i281.MethodChannel>(),
             ));
-    gh.factory<_i340.RhsNodeMapper>(
-        () => _i340.RhsNodeMapper(gh<_i1049.RhsNodeTypeMapper>()));
     gh.factory<_i361.SecureStorageDidProfileInfoDataSource>(() =>
         _i361.SecureStorageDidProfileInfoDataSource(
             gh<_i361.SecureDidProfileInfoStoreRefWrapper>()));
     gh.factoryAsync<_i540.CircuitsFilesDataSource>(() async =>
         _i540.CircuitsFilesDataSource(await getAsync<_i497.Directory>()));
+    gh.factory<_i66.DidProfileInfoRepositoryImpl>(
+        () => _i66.DidProfileInfoRepositoryImpl(
+              gh<_i361.SecureStorageDidProfileInfoDataSource>(),
+              gh<_i461.FiltersMapper>(),
+            ));
     gh.factory<_i352.CircuitsDownloadDataSource>(
         () => _i352.CircuitsDownloadDataSource(gh<_i361.Dio>()));
     gh.factory<_i232.SecureStorageProfilesStoreRefWrapper>(() =>
         _i232.SecureStorageProfilesStoreRefWrapper(
             gh<_i310.StoreRef<String, Map<String, Object?>>>(
                 instanceName: 'profilesStore')));
-    gh.factoryAsync<_i355.PackageInfoDataSource>(() async =>
-        _i355.PackageInfoDataSource(await getAsync<_i655.PackageInfo>()));
+    gh.singletonAsync<_i295.GetPackageNameUseCase>(() async =>
+        _i295.GetPackageNameUseCase(await getAsync<_i655.PackageInfo>()));
     gh.factory<_i698.Iden3commCredentialRepository>(() =>
         repositoriesModule.iden3commCredentialRepository(
             gh<_i910.Iden3commCredentialRepositoryImpl>()));
     gh.factory<_i938.DestinationPathDataSource>(
         () => _i938.DestinationPathDataSource(gh<_i938.CreatePathWrapper>()));
-    gh.factory<_i66.DidProfileInfoRepositoryImpl>(
-        () => _i66.DidProfileInfoRepositoryImpl(
-              gh<_i361.SecureStorageDidProfileInfoDataSource>(),
-              gh<_i461.FiltersMapper>(),
-              gh<_i435.DidProfileInfoInteractedDidFilterMapper>(),
-            ));
     gh.factory<_i41.LibPolygonIdCoreWrapper>(
         () => _i41.LibPolygonIdCoreWrapper(gh<_i961.PolygonIdCoreProof>()));
     gh.factory<_i42.StorageSMTDataSource>(
         () => _i42.StorageSMTDataSource(gh<_i42.IdentitySMTStoreRefWrapper>()));
-    gh.factory<_i882.LibPolygonIdCoreIden3commDataSource>(() =>
-        _i882.LibPolygonIdCoreIden3commDataSource(
-            gh<_i360.PolygonIdCoreIden3comm>()));
     gh.factory<_i232.SecureStorageProfilesDataSource>(() =>
         _i232.SecureStorageProfilesDataSource(
             gh<_i232.SecureStorageProfilesStoreRefWrapper>()));
+    gh.factory<_i575.SMTDataSource>(
+        () => _i575.SMTDataSource(gh<_i42.StorageSMTDataSource>()));
     gh.singleton<_i969.LocalClaimDataSource>(() => _i969.LocalClaimDataSource(
         gh<_i758.LibPolygonIdCoreCredentialDataSource>()));
     gh.factory<_i272.CacheCredentialDataSource>(() =>
@@ -699,6 +595,14 @@ extension GetItInjectableX on _i174.GetIt {
     gh.factory<_i525.KeyValueStoreRefWrapper>(() =>
         _i525.KeyValueStoreRefWrapper(gh<_i310.StoreRef<String, dynamic>>(
             instanceName: 'keyValueStore')));
+    gh.factoryAsync<_i548.InteractionRepositoryImpl>(
+        () async => _i548.InteractionRepositoryImpl(
+              gh<_i425.SecureStorageInteractionDataSource>(),
+              await getAsync<_i57.StorageInteractionDataSource>(),
+              gh<_i1026.InteractionMapper>(),
+              gh<_i461.FiltersMapper>(),
+              gh<_i267.StacktraceManager>(),
+            ));
     gh.factory<_i631.GetProofQueryContextUseCase>(
         () => _i631.GetProofQueryContextUseCase(
               gh<_i698.Iden3commCredentialRepository>(),
@@ -721,63 +625,19 @@ extension GetItInjectableX on _i174.GetIt {
     gh.factory<_i646.RemoveDidProfileInfoUseCase>(() =>
         _i646.RemoveDidProfileInfoUseCase(
             gh<_i258.DidProfileInfoRepository>()));
-    gh.factory<_i588.Iden3commRepositoryImpl>(
-        () => _i588.Iden3commRepositoryImpl(
-              gh<_i296.Iden3MessageDataSource>(),
-              gh<_i409.RemoteIden3commDataSource>(),
-              gh<_i882.LibPolygonIdCoreIden3commDataSource>(),
-              gh<_i22.AuthResponseMapper>(),
-              gh<_i1050.AuthProofMapper>(),
-              gh<_i178.GistMTProofMapper>(),
-              gh<_i599.QMapper>(),
-              gh<_i98.JWZMapper>(),
-              gh<_i897.GetIden3MessageUseCase>(),
-              gh<_i267.StacktraceManager>(),
-            ));
-    gh.factory<_i550.CredentialRepositoryImpl>(
-        () => _i550.CredentialRepositoryImpl(
-              gh<_i62.RemoteClaimDataSource>(),
-              gh<_i738.StorageClaimDataSource>(),
-              gh<_i969.LocalClaimDataSource>(),
-              gh<_i272.CacheCredentialDataSource>(),
-              gh<_i294.ClaimMapper>(),
-              gh<_i461.FiltersMapper>(),
-              gh<_i161.IdFilterMapper>(),
-              gh<_i267.StacktraceManager>(),
-            ));
-    gh.factory<_i575.SMTDataSource>(() => _i575.SMTDataSource(
-          gh<_i107.HexMapper>(),
-          gh<_i42.StorageSMTDataSource>(),
-        ));
     gh.factoryAsync<_i525.StorageKeyValueDataSource>(
         () async => _i525.StorageKeyValueDataSource(
               await getAsync<_i310.Database>(),
               gh<_i525.KeyValueStoreRefWrapper>(),
             ));
-    gh.factoryAsync<_i47.PackageInfoRepositoryImpl>(() async =>
-        _i47.PackageInfoRepositoryImpl(
-            await getAsync<_i355.PackageInfoDataSource>()));
-    gh.factory<_i309.CredentialRepository>(() => repositoriesModule
-        .credentialRepository(gh<_i550.CredentialRepositoryImpl>()));
-    gh.factory<_i53.GetClaimRevocationNonceUseCase>(() =>
-        _i53.GetClaimRevocationNonceUseCase(gh<_i309.CredentialRepository>()));
-    gh.factory<_i227.GetCredentialByIdUseCase>(
-        () => _i227.GetCredentialByIdUseCase(gh<_i309.CredentialRepository>()));
-    gh.factory<_i348.CacheCredentialUseCase>(
-        () => _i348.CacheCredentialUseCase(gh<_i309.CredentialRepository>()));
-    gh.factoryAsync<_i548.InteractionRepositoryImpl>(
-        () async => _i548.InteractionRepositoryImpl(
-              gh<_i425.SecureStorageInteractionDataSource>(),
-              await getAsync<_i57.StorageInteractionDataSource>(),
-              gh<_i1026.InteractionMapper>(),
-              gh<_i461.FiltersMapper>(),
-              gh<_i609.InteractionIdFilterMapper>(),
-              gh<_i267.StacktraceManager>(),
-            ));
     gh.factory<_i238.FetchSchemaUseCase>(() =>
         _i238.FetchSchemaUseCase(gh<_i698.Iden3commCredentialRepository>()));
     gh.factory<_i233.GetSchemasUseCase>(() =>
         _i233.GetSchemasUseCase(gh<_i698.Iden3commCredentialRepository>()));
+    gh.factory<_i328.SMTRepositoryImpl>(() => _i328.SMTRepositoryImpl(
+          gh<_i575.SMTDataSource>(),
+          gh<_i42.StorageSMTDataSource>(),
+        ));
     gh.factory<_i627.GetProofRequestsUseCase>(
         () => _i627.GetProofRequestsUseCase(
               gh<_i631.GetProofQueryContextUseCase>(),
@@ -786,66 +646,46 @@ extension GetItInjectableX on _i174.GetIt {
     gh.factoryAsync<_i1012.InteractionRepository>(() async =>
         repositoriesModule.interactionRepository(
             await getAsync<_i548.InteractionRepositoryImpl>()));
-    gh.factory<_i958.RemoveClaimsUseCase>(() => _i958.RemoveClaimsUseCase(
-          gh<_i309.CredentialRepository>(),
-          gh<_i267.StacktraceManager>(),
-        ));
-    gh.factory<_i158.GetCredentialByPartialIdUseCase>(
-        () => _i158.GetCredentialByPartialIdUseCase(
-              gh<_i309.CredentialRepository>(),
+    gh.factory<_i550.CredentialRepositoryImpl>(
+        () => _i550.CredentialRepositoryImpl(
+              gh<_i62.RemoteClaimDataSource>(),
+              gh<_i738.StorageClaimDataSource>(),
+              gh<_i969.LocalClaimDataSource>(),
+              gh<_i272.CacheCredentialDataSource>(),
+              gh<_i294.ClaimMapper>(),
+              gh<_i461.FiltersMapper>(),
               gh<_i267.StacktraceManager>(),
             ));
-    gh.factory<_i168.UpdateClaimUseCase>(() => _i168.UpdateClaimUseCase(
-          gh<_i309.CredentialRepository>(),
-          gh<_i267.StacktraceManager>(),
-        ));
-    gh.factory<_i635.SaveClaimsUseCase>(() => _i635.SaveClaimsUseCase(
-          gh<_i309.CredentialRepository>(),
-          gh<_i267.StacktraceManager>(),
-        ));
-    gh.factory<_i503.RemoveAllClaimsUseCase>(() => _i503.RemoveAllClaimsUseCase(
-          gh<_i309.CredentialRepository>(),
-          gh<_i267.StacktraceManager>(),
-        ));
-    gh.factoryAsync<_i141.PackageInfoRepository>(() async =>
-        repositoriesModule.packageInfoRepository(
-            await getAsync<_i47.PackageInfoRepositoryImpl>()));
-    gh.factory<_i392.GetAuthClaimUseCase>(() => _i392.GetAuthClaimUseCase(
-          gh<_i309.CredentialRepository>(),
-          gh<_i267.StacktraceManager>(),
-        ));
-    gh.singletonAsync<_i1049.ConfigRepositoryImpl>(() async =>
-        _i1049.ConfigRepositoryImpl(
+    gh.factory<_i588.Iden3commRepositoryImpl>(
+        () => _i588.Iden3commRepositoryImpl(
+              gh<_i296.Iden3MessageDataSource>(),
+              gh<_i409.RemoteIden3commDataSource>(),
+              gh<_i41.LibPolygonIdCoreProofDataSource>(),
+              gh<_i599.QMapper>(),
+              gh<_i98.JWZMapper>(),
+              gh<_i897.GetIden3MessageUseCase>(),
+              gh<_i267.StacktraceManager>(),
+            ));
+    gh.singletonAsync<_i415.ConfigRepository>(() async =>
+        _i415.ConfigRepository(
             await getAsync<_i525.StorageKeyValueDataSource>()));
-    gh.factory<_i328.SMTRepositoryImpl>(() => _i328.SMTRepositoryImpl(
-          gh<_i575.SMTDataSource>(),
-          gh<_i42.StorageSMTDataSource>(),
-          gh<_i443.TreeTypeMapper>(),
-          gh<_i68.TreeStateMapper>(),
-        ));
+    gh.factory<_i946.SMTRepository>(
+        () => repositoriesModule.smtRepository(gh<_i328.SMTRepositoryImpl>()));
     gh.factory<_i88.Iden3commRepository>(() => repositoriesModule
         .iden3commRepository(gh<_i588.Iden3commRepositoryImpl>()));
     gh.factory<_i359.CleanSchemaCacheUseCase>(
         () => _i359.CleanSchemaCacheUseCase(gh<_i88.Iden3commRepository>()));
-    gh.factoryAsync<_i191.ListenAndStoreNotificationUseCase>(() async =>
-        _i191.ListenAndStoreNotificationUseCase(
-            await getAsync<_i1012.InteractionRepository>()));
-    gh.factory<_i734.GetAuthChallengeUseCase>(
-        () => _i734.GetAuthChallengeUseCase(
-              gh<_i88.Iden3commRepository>(),
-              gh<_i267.StacktraceManager>(),
-            ));
-    gh.factory<_i249.GetJWZUseCase>(() => _i249.GetJWZUseCase(
-          gh<_i88.Iden3commRepository>(),
-          gh<_i267.StacktraceManager>(),
-        ));
-    gh.factoryAsync<_i295.GetPackageNameUseCase>(() async =>
-        _i295.GetPackageNameUseCase(
-            await getAsync<_i141.PackageInfoRepository>()));
-    gh.factoryAsync<_i415.ConfigRepository>(() async => repositoriesModule
-        .configRepository(await getAsync<_i1049.ConfigRepositoryImpl>()));
-    gh.factory<_i946.SMTRepository>(
-        () => repositoriesModule.smtRepository(gh<_i328.SMTRepositoryImpl>()));
+    gh.factory<_i309.CredentialRepository>(() => repositoriesModule
+        .credentialRepository(gh<_i550.CredentialRepositoryImpl>()));
+    gh.factory<_i53.GetClaimRevocationNonceUseCase>(() =>
+        _i53.GetClaimRevocationNonceUseCase(gh<_i309.CredentialRepository>()));
+    gh.factory<_i227.GetCredentialByIdUseCase>(
+        () => _i227.GetCredentialByIdUseCase(gh<_i309.CredentialRepository>()));
+    gh.factory<_i158.GetCredentialByPartialIdUseCase>(() =>
+        _i158.GetCredentialByPartialIdUseCase(
+            gh<_i309.CredentialRepository>()));
+    gh.factory<_i348.CacheCredentialUseCase>(
+        () => _i348.CacheCredentialUseCase(gh<_i309.CredentialRepository>()));
     gh.factoryAsync<_i438.SetSelectedChainUseCase>(() async =>
         _i438.SetSelectedChainUseCase(
             await getAsync<_i415.ConfigRepository>()));
@@ -860,6 +700,39 @@ extension GetItInjectableX on _i174.GetIt {
               gh<_i348.CacheCredentialUseCase>(),
               gh<_i267.StacktraceManager>(),
               await getAsync<_i626.GetEnvUseCase>(),
+            ));
+    gh.factoryAsync<_i581.ProofRepositoryImpl>(
+        () async => _i581.ProofRepositoryImpl(
+              gh<_i1039.WitnessDataSource>(),
+              gh<_i502.ProverLibDataSource>(),
+              gh<_i41.LibPolygonIdCoreProofDataSource>(),
+              gh<_i694.GistMTProofDataSource>(),
+              gh<_i1021.ProofCircuitDataSource>(),
+              gh<_i22.LocalContractFilesDataSource>(),
+              gh<_i352.CircuitsDownloadDataSource>(),
+              gh<_i294.ClaimMapper>(),
+              await getAsync<_i540.CircuitsFilesDataSource>(),
+              await getAsync<_i626.GetEnvUseCase>(),
+              gh<_i267.StacktraceManager>(),
+            ));
+    gh.factory<_i734.GetAuthChallengeUseCase>(
+        () => _i734.GetAuthChallengeUseCase(
+              gh<_i88.Iden3commRepository>(),
+              gh<_i267.StacktraceManager>(),
+            ));
+    gh.factory<_i249.GetJWZUseCase>(() => _i249.GetJWZUseCase(
+          gh<_i88.Iden3commRepository>(),
+          gh<_i267.StacktraceManager>(),
+        ));
+    gh.factoryAsync<_i1031.AddInteractionUseCase>(
+        () async => _i1031.AddInteractionUseCase(
+              await getAsync<_i1012.InteractionRepository>(),
+              gh<_i267.StacktraceManager>(),
+            ));
+    gh.factoryAsync<_i484.GetInteractionsUseCase>(
+        () async => _i484.GetInteractionsUseCase(
+              await getAsync<_i1012.InteractionRepository>(),
+              gh<_i267.StacktraceManager>(),
             ));
     gh.factoryAsync<_i737.GetSelectedChainUseCase>(
         () async => _i737.GetSelectedChainUseCase(
@@ -879,6 +752,30 @@ extension GetItInjectableX on _i174.GetIt {
           await getAsync<_i737.GetSelectedChainUseCase>(),
           gh<_i267.StacktraceManager>(),
         ));
+    gh.factory<_i958.RemoveClaimsUseCase>(() => _i958.RemoveClaimsUseCase(
+          gh<_i309.CredentialRepository>(),
+          gh<_i267.StacktraceManager>(),
+        ));
+    gh.factory<_i660.GetNonRevProofUseCase>(() => _i660.GetNonRevProofUseCase(
+          gh<_i309.CredentialRepository>(),
+          gh<_i267.StacktraceManager>(),
+        ));
+    gh.factory<_i657.GetClaimsUseCase>(() => _i657.GetClaimsUseCase(
+          gh<_i309.CredentialRepository>(),
+          gh<_i267.StacktraceManager>(),
+        ));
+    gh.factory<_i168.UpdateClaimUseCase>(() => _i168.UpdateClaimUseCase(
+          gh<_i309.CredentialRepository>(),
+          gh<_i267.StacktraceManager>(),
+        ));
+    gh.factory<_i635.SaveClaimsUseCase>(() => _i635.SaveClaimsUseCase(
+          gh<_i309.CredentialRepository>(),
+          gh<_i267.StacktraceManager>(),
+        ));
+    gh.factory<_i503.RemoveAllClaimsUseCase>(() => _i503.RemoveAllClaimsUseCase(
+          gh<_i309.CredentialRepository>(),
+          gh<_i267.StacktraceManager>(),
+        ));
     gh.factoryAsync<_i1054.FetchOnchainClaimUseCase>(
         () async => _i1054.FetchOnchainClaimUseCase(
               await getAsync<_i737.GetSelectedChainUseCase>(),
@@ -889,42 +786,10 @@ extension GetItInjectableX on _i174.GetIt {
               gh<_i294.ClaimMapper>(),
               gh<_i267.StacktraceManager>(),
             ));
-    gh.factoryAsync<_i581.ProofRepositoryImpl>(
-        () async => _i581.ProofRepositoryImpl(
-              gh<_i1039.WitnessDataSource>(),
-              gh<_i502.ProverLibDataSource>(),
-              gh<_i41.LibPolygonIdCoreProofDataSource>(),
-              gh<_i694.GistMTProofDataSource>(),
-              gh<_i1021.ProofCircuitDataSource>(),
-              gh<_i839.RemoteIdentityDataSource>(),
-              gh<_i22.LocalContractFilesDataSource>(),
-              gh<_i352.CircuitsDownloadDataSource>(),
-              await getAsync<_i873.RPCDataSource>(),
-              gh<_i512.CircuitTypeMapper>(),
-              gh<_i360.ZKProofMapper>(),
-              gh<_i294.ClaimMapper>(),
-              gh<_i162.RevocationStatusMapper>(),
-              gh<_i1050.AuthProofMapper>(),
-              await getAsync<_i540.CircuitsFilesDataSource>(),
-              await getAsync<_i626.GetEnvUseCase>(),
-              gh<_i267.StacktraceManager>(),
-            ));
-    gh.factoryAsync<_i393.IdentityRepositoryImpl>(
-        () async => _i393.IdentityRepositoryImpl(
-              gh<_i383.WalletDataSource>(),
-              gh<_i839.RemoteIdentityDataSource>(),
-              await getAsync<_i995.StorageIdentityDataSource>(),
-              await getAsync<_i873.RPCDataSource>(),
-              gh<_i22.LocalContractFilesDataSource>(),
-              gh<_i136.LibPolygonIdCoreIdentityDataSource>(),
-              gh<_i200.EncryptionDbDataSource>(),
-              gh<_i938.DestinationPathDataSource>(),
-              gh<_i107.HexMapper>(),
-              gh<_i526.PrivateKeyMapper>(),
-              gh<_i340.RhsNodeMapper>(),
-              gh<_i720.StateIdentifierMapper>(),
-              gh<_i232.SecureStorageProfilesDataSource>(),
-            ));
+    gh.factory<_i392.GetAuthClaimUseCase>(() => _i392.GetAuthClaimUseCase(
+          gh<_i309.CredentialRepository>(),
+          gh<_i267.StacktraceManager>(),
+        ));
     gh.factoryAsync<_i341.ProofRepository>(() async => repositoriesModule
         .proofRepository(await getAsync<_i581.ProofRepositoryImpl>()));
     gh.factoryAsync<_i660.LoadCircuitUseCase>(
@@ -942,29 +807,19 @@ extension GetItInjectableX on _i174.GetIt {
               await getAsync<_i310.ProveUseCase>(),
               gh<_i267.StacktraceManager>(),
             ));
-    gh.factoryAsync<_i26.IdentityRepository>(() async => repositoriesModule
-        .identityRepository(await getAsync<_i393.IdentityRepositoryImpl>()));
-    gh.factoryAsync<_i449.SignMessageUseCase>(() async =>
-        _i449.SignMessageUseCase(await getAsync<_i26.IdentityRepository>()));
-    gh.factoryAsync<_i172.FetchStateRootsUseCase>(() async =>
-        _i172.FetchStateRootsUseCase(
-            await getAsync<_i26.IdentityRepository>()));
-    gh.factoryAsync<_i78.GetDidUseCase>(() async =>
-        _i78.GetDidUseCase(await getAsync<_i26.IdentityRepository>()));
-    gh.factoryAsync<_i665.GetPrivateKeyUseCase>(
-        () async => _i665.GetPrivateKeyUseCase(
-              await getAsync<_i26.IdentityRepository>(),
-              gh<_i267.StacktraceManager>(),
-            ));
-    gh.factoryAsync<_i348.GetIdentitiesUseCase>(
-        () async => _i348.GetIdentitiesUseCase(
-              await getAsync<_i26.IdentityRepository>(),
-              gh<_i267.StacktraceManager>(),
-            ));
-    gh.factoryAsync<_i166.GetPublicKeyUseCase>(
-        () async => _i166.GetPublicKeyUseCase(
-              await getAsync<_i26.IdentityRepository>(),
-              gh<_i267.StacktraceManager>(),
+    gh.factoryAsync<_i393.IdentityRepositoryImpl>(
+        () async => _i393.IdentityRepositoryImpl(
+              gh<_i383.WalletDataSource>(),
+              gh<_i839.RemoteIdentityDataSource>(),
+              await getAsync<_i995.StorageIdentityDataSource>(),
+              await getAsync<_i873.RPCDataSource>(),
+              gh<_i22.LocalContractFilesDataSource>(),
+              gh<_i136.LibPolygonIdCoreIdentityDataSource>(),
+              gh<_i200.EncryptionDbDataSource>(),
+              gh<_i938.DestinationPathDataSource>(),
+              gh<_i526.PrivateKeyMapper>(),
+              gh<_i720.StateIdentifierMapper>(),
+              gh<_i232.SecureStorageProfilesDataSource>(),
             ));
     gh.factoryAsync<_i735.IsProofCircuitSupportedUseCase>(() async =>
         _i735.IsProofCircuitSupportedUseCase(
@@ -977,19 +832,12 @@ extension GetItInjectableX on _i174.GetIt {
     gh.factoryAsync<_i394.CancelDownloadCircuitsUseCase>(() async =>
         _i394.CancelDownloadCircuitsUseCase(
             await getAsync<_i341.ProofRepository>()));
-    gh.factoryAsync<_i1042.GetGenesisStateUseCase>(
-        () async => _i1042.GetGenesisStateUseCase(
-              await getAsync<_i26.IdentityRepository>(),
-              gh<_i946.SMTRepository>(),
-              gh<_i392.GetAuthClaimUseCase>(),
-              gh<_i267.StacktraceManager>(),
-            ));
-    gh.factoryAsync<_i798.CreateIdentityStateUseCase>(
-        () async => _i798.CreateIdentityStateUseCase(
-              await getAsync<_i26.IdentityRepository>(),
-              gh<_i946.SMTRepository>(),
-              gh<_i392.GetAuthClaimUseCase>(),
-              await getAsync<_i166.GetPublicKeyUseCase>(),
+    gh.factoryAsync<_i347.GetIden3commClaimsUseCase>(
+        () async => _i347.GetIden3commClaimsUseCase(
+              gh<_i698.Iden3commCredentialRepository>(),
+              gh<_i657.GetClaimsUseCase>(),
+              await getAsync<_i735.IsProofCircuitSupportedUseCase>(),
+              gh<_i627.GetProofRequestsUseCase>(),
               gh<_i267.StacktraceManager>(),
             ));
     gh.factoryAsync<_i445.Proof>(() async => _i445.Proof(
@@ -1006,19 +854,56 @@ extension GetItInjectableX on _i174.GetIt {
               await getAsync<_i735.IsProofCircuitSupportedUseCase>(),
               gh<_i627.GetProofRequestsUseCase>(),
             ));
+    gh.factoryAsync<_i369.GetIden3commClaimsRevNonceUseCase>(
+        () async => _i369.GetIden3commClaimsRevNonceUseCase(
+              gh<_i698.Iden3commCredentialRepository>(),
+              gh<_i657.GetClaimsUseCase>(),
+              gh<_i53.GetClaimRevocationNonceUseCase>(),
+              await getAsync<_i735.IsProofCircuitSupportedUseCase>(),
+              gh<_i627.GetProofRequestsUseCase>(),
+            ));
+    gh.factoryAsync<_i26.IdentityRepository>(() async => repositoriesModule
+        .identityRepository(await getAsync<_i393.IdentityRepositoryImpl>()));
+    gh.factoryAsync<_i449.SignMessageUseCase>(() async =>
+        _i449.SignMessageUseCase(await getAsync<_i26.IdentityRepository>()));
+    gh.factoryAsync<_i172.FetchStateRootsUseCase>(() async =>
+        _i172.FetchStateRootsUseCase(
+            await getAsync<_i26.IdentityRepository>()));
+    gh.factoryAsync<_i665.GetPrivateKeyUseCase>(
+        () async => _i665.GetPrivateKeyUseCase(
+              await getAsync<_i26.IdentityRepository>(),
+              gh<_i267.StacktraceManager>(),
+            ));
+    gh.factoryAsync<_i348.GetIdentitiesUseCase>(
+        () async => _i348.GetIdentitiesUseCase(
+              await getAsync<_i26.IdentityRepository>(),
+              gh<_i267.StacktraceManager>(),
+            ));
+    gh.factoryAsync<_i166.GetPublicKeysUseCase>(
+        () async => _i166.GetPublicKeysUseCase(
+              await getAsync<_i26.IdentityRepository>(),
+              gh<_i267.StacktraceManager>(),
+            ));
+    gh.factoryAsync<_i1042.GetGenesisStateUseCase>(
+        () async => _i1042.GetGenesisStateUseCase(
+              await getAsync<_i26.IdentityRepository>(),
+              gh<_i946.SMTRepository>(),
+              gh<_i392.GetAuthClaimUseCase>(),
+              gh<_i267.StacktraceManager>(),
+            ));
     gh.factoryAsync<_i344.GetGistMTProofUseCase>(
         () async => _i344.GetGistMTProofUseCase(
               await getAsync<_i341.ProofRepository>(),
               await getAsync<_i26.IdentityRepository>(),
               await getAsync<_i737.GetSelectedChainUseCase>(),
-              await getAsync<_i78.GetDidUseCase>(),
+              gh<_i78.GetDidUseCase>(),
               gh<_i267.StacktraceManager>(),
             ));
     gh.factoryAsync<_i484.FetchIdentityStateUseCase>(
         () async => _i484.FetchIdentityStateUseCase(
               await getAsync<_i26.IdentityRepository>(),
               await getAsync<_i737.GetSelectedChainUseCase>(),
-              await getAsync<_i78.GetDidUseCase>(),
+              gh<_i78.GetDidUseCase>(),
               gh<_i267.StacktraceManager>(),
             ));
     gh.factoryAsync<_i732.GetDidIdentifierUseCase>(
@@ -1028,10 +913,12 @@ extension GetItInjectableX on _i174.GetIt {
               await getAsync<_i1042.GetGenesisStateUseCase>(),
               gh<_i267.StacktraceManager>(),
             ));
-    gh.factoryAsync<_i675.GetCurrentEnvDidIdentifierUseCase>(
-        () async => _i675.GetCurrentEnvDidIdentifierUseCase(
-              await getAsync<_i737.GetSelectedChainUseCase>(),
-              await getAsync<_i732.GetDidIdentifierUseCase>(),
+    gh.factoryAsync<_i798.CreateIdentityStateUseCase>(
+        () async => _i798.CreateIdentityStateUseCase(
+              await getAsync<_i26.IdentityRepository>(),
+              gh<_i946.SMTRepository>(),
+              gh<_i392.GetAuthClaimUseCase>(),
+              await getAsync<_i166.GetPublicKeysUseCase>(),
               gh<_i267.StacktraceManager>(),
             ));
     gh.factoryAsync<_i505.CheckProfileAndDidCurrentEnvUseCase>(
@@ -1039,14 +926,24 @@ extension GetItInjectableX on _i174.GetIt {
               gh<_i192.CheckProfileValidityUseCase>(),
               await getAsync<_i737.GetSelectedChainUseCase>(),
               await getAsync<_i732.GetDidIdentifierUseCase>(),
-              await getAsync<_i166.GetPublicKeyUseCase>(),
               gh<_i267.StacktraceManager>(),
             ));
-    gh.factoryAsync<_i660.GetNonRevProofUseCase>(
-        () async => _i660.GetNonRevProofUseCase(
+    gh.factoryAsync<_i146.FetchOnchainClaimsUseCase>(
+        () async => _i146.FetchOnchainClaimsUseCase(
+              await getAsync<_i1054.FetchOnchainClaimUseCase>(),
+              await getAsync<_i505.CheckProfileAndDidCurrentEnvUseCase>(),
+              await getAsync<_i626.GetEnvUseCase>(),
+              await getAsync<_i737.GetSelectedChainUseCase>(),
+              gh<_i78.GetDidUseCase>(),
+              gh<_i22.LocalContractFilesDataSource>(),
               await getAsync<_i26.IdentityRepository>(),
-              gh<_i309.CredentialRepository>(),
-              await getAsync<_i484.FetchIdentityStateUseCase>(),
+              gh<_i258.DidProfileInfoRepository>(),
+              gh<_i267.StacktraceManager>(),
+            ));
+    gh.factoryAsync<_i675.GetCurrentEnvDidIdentifierUseCase>(
+        () async => _i675.GetCurrentEnvDidIdentifierUseCase(
+              await getAsync<_i737.GetSelectedChainUseCase>(),
+              await getAsync<_i732.GetDidIdentifierUseCase>(),
               gh<_i267.StacktraceManager>(),
             ));
     gh.factoryAsync<_i40.GenerateNonRevProofUseCase>(
@@ -1056,10 +953,16 @@ extension GetItInjectableX on _i174.GetIt {
               await getAsync<_i484.FetchIdentityStateUseCase>(),
               gh<_i267.StacktraceManager>(),
             ));
+    gh.factoryAsync<_i266.CreateProfilesUseCase>(
+        () async => _i266.CreateProfilesUseCase(
+              await getAsync<_i166.GetPublicKeysUseCase>(),
+              await getAsync<_i675.GetCurrentEnvDidIdentifierUseCase>(),
+              gh<_i267.StacktraceManager>(),
+            ));
     gh.factoryAsync<_i743.GetIdentityUseCase>(
         () async => _i743.GetIdentityUseCase(
               await getAsync<_i26.IdentityRepository>(),
-              await getAsync<_i78.GetDidUseCase>(),
+              gh<_i78.GetDidUseCase>(),
               await getAsync<_i732.GetDidIdentifierUseCase>(),
               gh<_i267.StacktraceManager>(),
             ));
@@ -1077,11 +980,7 @@ extension GetItInjectableX on _i174.GetIt {
             ));
     gh.factoryAsync<_i845.CreateIdentityUseCase>(
         () async => _i845.CreateIdentityUseCase(
-              await getAsync<_i675.GetCurrentEnvDidIdentifierUseCase>(),
-              gh<_i267.StacktraceManager>(),
-            ));
-    gh.factoryAsync<_i266.CreateProfilesUseCase>(
-        () async => _i266.CreateProfilesUseCase(
+              await getAsync<_i166.GetPublicKeysUseCase>(),
               await getAsync<_i675.GetCurrentEnvDidIdentifierUseCase>(),
               gh<_i267.StacktraceManager>(),
             ));
@@ -1094,52 +993,22 @@ extension GetItInjectableX on _i174.GetIt {
               await getAsync<_i310.ProveUseCase>(),
               gh<_i267.StacktraceManager>(),
             ));
+    gh.factoryAsync<_i816.UpdateIdentityUseCase>(
+        () async => _i816.UpdateIdentityUseCase(
+              await getAsync<_i26.IdentityRepository>(),
+              await getAsync<_i743.GetIdentityUseCase>(),
+            ));
     gh.factoryAsync<_i133.BackupIdentityUseCase>(
         () async => _i133.BackupIdentityUseCase(
               await getAsync<_i743.GetIdentityUseCase>(),
               await getAsync<_i26.IdentityRepository>(),
-              await getAsync<_i675.GetCurrentEnvDidIdentifierUseCase>(),
-              gh<_i267.StacktraceManager>(),
-            ));
-    gh.factoryAsync<_i816.UpdateIdentityUseCase>(
-        () async => _i816.UpdateIdentityUseCase(
-              await getAsync<_i26.IdentityRepository>(),
-              await getAsync<_i845.CreateIdentityUseCase>(),
-              await getAsync<_i743.GetIdentityUseCase>(),
-            ));
-    gh.factoryAsync<_i657.GetClaimsUseCase>(() async => _i657.GetClaimsUseCase(
-          gh<_i309.CredentialRepository>(),
-          await getAsync<_i675.GetCurrentEnvDidIdentifierUseCase>(),
-          gh<_i267.StacktraceManager>(),
-        ));
-    gh.factoryAsync<_i146.FetchOnchainClaimsUseCase>(
-        () async => _i146.FetchOnchainClaimsUseCase(
-              await getAsync<_i1054.FetchOnchainClaimUseCase>(),
-              await getAsync<_i505.CheckProfileAndDidCurrentEnvUseCase>(),
-              await getAsync<_i626.GetEnvUseCase>(),
-              await getAsync<_i737.GetSelectedChainUseCase>(),
-              await getAsync<_i732.GetDidIdentifierUseCase>(),
-              await getAsync<_i78.GetDidUseCase>(),
-              gh<_i635.SaveClaimsUseCase>(),
-              gh<_i348.CacheCredentialUseCase>(),
-              await getAsync<_i166.GetPublicKeyUseCase>(),
-              gh<_i22.LocalContractFilesDataSource>(),
-              await getAsync<_i26.IdentityRepository>(),
-              gh<_i258.DidProfileInfoRepository>(),
-              gh<_i267.StacktraceManager>(),
-            ));
-    gh.factoryAsync<_i548.CheckIdentityValidityUseCase>(
-        () async => _i548.CheckIdentityValidityUseCase(
-              await getAsync<_i665.GetPrivateKeyUseCase>(),
-              await getAsync<_i166.GetPublicKeyUseCase>(),
-              await getAsync<_i675.GetCurrentEnvDidIdentifierUseCase>(),
               gh<_i267.StacktraceManager>(),
             ));
     gh.factoryAsync<_i610.GetClaimRevocationStatusUseCase>(
         () async => _i610.GetClaimRevocationStatusUseCase(
               gh<_i309.CredentialRepository>(),
               await getAsync<_i40.GenerateNonRevProofUseCase>(),
-              await getAsync<_i660.GetNonRevProofUseCase>(),
+              gh<_i660.GetNonRevProofUseCase>(),
               gh<_i267.StacktraceManager>(),
             ));
     gh.factoryAsync<_i561.GetProfilesUseCase>(
@@ -1153,6 +1022,13 @@ extension GetItInjectableX on _i174.GetIt {
               await getAsync<_i1012.InteractionRepository>(),
               await getAsync<_i743.GetIdentityUseCase>(),
             ));
+    gh.factoryAsync<_i548.CheckIdentityValidityUseCase>(
+        () async => _i548.CheckIdentityValidityUseCase(
+              await getAsync<_i665.GetPrivateKeyUseCase>(),
+              await getAsync<_i166.GetPublicKeysUseCase>(),
+              await getAsync<_i675.GetCurrentEnvDidIdentifierUseCase>(),
+              gh<_i267.StacktraceManager>(),
+            ));
     gh.factoryAsync<_i102.FetchAndSaveClaimsUseCase>(
         () async => _i102.FetchAndSaveClaimsUseCase(
               gh<_i698.Iden3commCredentialRepository>(),
@@ -1161,32 +1037,13 @@ extension GetItInjectableX on _i174.GetIt {
               await getAsync<_i626.GetEnvUseCase>(),
               await getAsync<_i737.GetSelectedChainUseCase>(),
               await getAsync<_i732.GetDidIdentifierUseCase>(),
-              await getAsync<_i78.GetDidUseCase>(),
+              gh<_i78.GetDidUseCase>(),
               gh<_i968.GetFetchRequestsUseCase>(),
               await getAsync<_i871.GetAuthTokenUseCase>(),
               gh<_i635.SaveClaimsUseCase>(),
               gh<_i348.CacheCredentialUseCase>(),
               gh<_i22.LocalContractFilesDataSource>(),
               await getAsync<_i26.IdentityRepository>(),
-              await getAsync<_i1012.InteractionRepository>(),
-              gh<_i267.StacktraceManager>(),
-            ));
-    gh.factoryAsync<_i561.AddIdentityUseCase>(
-        () async => _i561.AddIdentityUseCase(
-              await getAsync<_i26.IdentityRepository>(),
-              await getAsync<_i845.CreateIdentityUseCase>(),
-              await getAsync<_i798.CreateIdentityStateUseCase>(),
-              await getAsync<_i166.GetPublicKeyUseCase>(),
-              gh<_i267.StacktraceManager>(),
-            ));
-    gh.factoryAsync<_i1050.AddProfileUseCase>(
-        () async => _i1050.AddProfileUseCase(
-              await getAsync<_i743.GetIdentityUseCase>(),
-              await getAsync<_i816.UpdateIdentityUseCase>(),
-              await getAsync<_i505.CheckProfileAndDidCurrentEnvUseCase>(),
-              await getAsync<_i266.CreateProfilesUseCase>(),
-              await getAsync<_i166.GetPublicKeyUseCase>(),
-              gh<_i136.LibPolygonIdCoreIdentityDataSource>(),
               gh<_i267.StacktraceManager>(),
             ));
     gh.factoryAsync<_i829.RemoveProfileUseCase>(
@@ -1197,24 +1054,14 @@ extension GetItInjectableX on _i174.GetIt {
               await getAsync<_i266.CreateProfilesUseCase>(),
               gh<_i1009.RemoveIdentityStateUseCase>(),
               gh<_i503.RemoveAllClaimsUseCase>(),
-              await getAsync<_i166.GetPublicKeyUseCase>(),
               gh<_i267.StacktraceManager>(),
             ));
-    gh.factoryAsync<_i143.RefreshCredentialUseCase>(
-        () async => _i143.RefreshCredentialUseCase(
-              gh<_i309.CredentialRepository>(),
-              gh<_i267.StacktraceManager>(),
+    gh.factoryAsync<_i989.UpdateInteractionUseCase>(
+        () async => _i989.UpdateInteractionUseCase(
+              await getAsync<_i1012.InteractionRepository>(),
+              gh<_i192.CheckProfileValidityUseCase>(),
               await getAsync<_i743.GetIdentityUseCase>(),
-              await getAsync<_i871.GetAuthTokenUseCase>(),
-              gh<_i698.Iden3commCredentialRepository>(),
-              gh<_i958.RemoveClaimsUseCase>(),
-              gh<_i635.SaveClaimsUseCase>(),
-            ));
-    gh.factoryAsync<_i279.AddNewIdentityUseCase>(
-        () async => _i279.AddNewIdentityUseCase(
-              await getAsync<_i26.IdentityRepository>(),
-              await getAsync<_i561.AddIdentityUseCase>(),
-              gh<_i267.StacktraceManager>(),
+              await getAsync<_i1031.AddInteractionUseCase>(),
             ));
     gh.factoryAsync<_i709.FetchCredentialsUseCase>(
         () async => _i709.FetchCredentialsUseCase(
@@ -1226,23 +1073,9 @@ extension GetItInjectableX on _i174.GetIt {
               gh<_i267.StacktraceManager>(),
               await getAsync<_i26.IdentityRepository>(),
               gh<_i22.LocalContractFilesDataSource>(),
-              await getAsync<_i78.GetDidUseCase>(),
+              gh<_i78.GetDidUseCase>(),
               await getAsync<_i1054.FetchOnchainClaimUseCase>(),
               await getAsync<_i626.GetEnvUseCase>(),
-            ));
-    gh.factoryAsync<_i1031.AddInteractionUseCase>(
-        () async => _i1031.AddInteractionUseCase(
-              await getAsync<_i1012.InteractionRepository>(),
-              gh<_i192.CheckProfileValidityUseCase>(),
-              await getAsync<_i743.GetIdentityUseCase>(),
-              gh<_i267.StacktraceManager>(),
-            ));
-    gh.factoryAsync<_i484.GetInteractionsUseCase>(
-        () async => _i484.GetInteractionsUseCase(
-              await getAsync<_i1012.InteractionRepository>(),
-              gh<_i192.CheckProfileValidityUseCase>(),
-              await getAsync<_i743.GetIdentityUseCase>(),
-              gh<_i267.StacktraceManager>(),
             ));
     gh.factoryAsync<_i668.RemoveIdentityUseCase>(
         () async => _i668.RemoveIdentityUseCase(
@@ -1254,13 +1087,14 @@ extension GetItInjectableX on _i174.GetIt {
               await getAsync<_i505.CheckProfileAndDidCurrentEnvUseCase>(),
               gh<_i267.StacktraceManager>(),
             ));
-    gh.factoryAsync<_i369.GetIden3commClaimsRevNonceUseCase>(
-        () async => _i369.GetIden3commClaimsRevNonceUseCase(
-              gh<_i698.Iden3commCredentialRepository>(),
-              await getAsync<_i657.GetClaimsUseCase>(),
-              gh<_i53.GetClaimRevocationNonceUseCase>(),
-              await getAsync<_i735.IsProofCircuitSupportedUseCase>(),
-              gh<_i627.GetProofRequestsUseCase>(),
+    gh.factoryAsync<_i1050.AddProfileUseCase>(
+        () async => _i1050.AddProfileUseCase(
+              await getAsync<_i743.GetIdentityUseCase>(),
+              await getAsync<_i816.UpdateIdentityUseCase>(),
+              await getAsync<_i505.CheckProfileAndDidCurrentEnvUseCase>(),
+              await getAsync<_i266.CreateProfilesUseCase>(),
+              gh<_i136.LibPolygonIdCoreIdentityDataSource>(),
+              gh<_i267.StacktraceManager>(),
             ));
     gh.factoryAsync<_i340.GenerateIden3commProofUseCase>(
         () async => _i340.GenerateIden3commProofUseCase(
@@ -1271,9 +1105,25 @@ extension GetItInjectableX on _i174.GetIt {
               await getAsync<_i743.GetIdentityUseCase>(),
               gh<_i392.GetAuthClaimUseCase>(),
               await getAsync<_i344.GetGistMTProofUseCase>(),
-              await getAsync<_i78.GetDidUseCase>(),
+              gh<_i78.GetDidUseCase>(),
               await getAsync<_i449.SignMessageUseCase>(),
               gh<_i754.GetLatestStateUseCase>(),
+              gh<_i267.StacktraceManager>(),
+            ));
+    gh.factoryAsync<_i143.RefreshCredentialUseCase>(
+        () async => _i143.RefreshCredentialUseCase(
+              gh<_i267.StacktraceManager>(),
+              await getAsync<_i743.GetIdentityUseCase>(),
+              await getAsync<_i871.GetAuthTokenUseCase>(),
+              gh<_i698.Iden3commCredentialRepository>(),
+              gh<_i958.RemoveClaimsUseCase>(),
+              gh<_i635.SaveClaimsUseCase>(),
+            ));
+    gh.factoryAsync<_i561.AddIdentityUseCase>(
+        () async => _i561.AddIdentityUseCase(
+              await getAsync<_i26.IdentityRepository>(),
+              await getAsync<_i845.CreateIdentityUseCase>(),
+              await getAsync<_i798.CreateIdentityStateUseCase>(),
               gh<_i267.StacktraceManager>(),
             ));
     gh.factoryAsync<_i657.RestoreProfilesUseCase>(
@@ -1283,7 +1133,7 @@ extension GetItInjectableX on _i174.GetIt {
             ));
     gh.factoryAsync<_i501.Credential>(() async => _i501.Credential(
           gh<_i635.SaveClaimsUseCase>(),
-          await getAsync<_i657.GetClaimsUseCase>(),
+          gh<_i657.GetClaimsUseCase>(),
           gh<_i958.RemoveClaimsUseCase>(),
           await getAsync<_i610.GetClaimRevocationStatusUseCase>(),
           gh<_i168.UpdateClaimUseCase>(),
@@ -1302,18 +1152,6 @@ extension GetItInjectableX on _i174.GetIt {
               await getAsync<_i675.GetCurrentEnvDidIdentifierUseCase>(),
               await getAsync<_i657.RestoreProfilesUseCase>(),
             ));
-    gh.factoryAsync<_i347.GetIden3commClaimsUseCase>(
-        () async => _i347.GetIden3commClaimsUseCase(
-              gh<_i698.Iden3commCredentialRepository>(),
-              await getAsync<_i657.GetClaimsUseCase>(),
-              await getAsync<_i610.GetClaimRevocationStatusUseCase>(),
-              gh<_i53.GetClaimRevocationNonceUseCase>(),
-              gh<_i168.UpdateClaimUseCase>(),
-              await getAsync<_i735.IsProofCircuitSupportedUseCase>(),
-              gh<_i627.GetProofRequestsUseCase>(),
-              gh<_i512.CircuitTypeMapper>(),
-              gh<_i267.StacktraceManager>(),
-            ));
     gh.factoryAsync<_i412.GetIden3commProofsUseCase>(
         () async => _i412.GetIden3commProofsUseCase(
               await getAsync<_i341.ProofRepository>(),
@@ -1324,11 +1162,13 @@ extension GetItInjectableX on _i174.GetIt {
               await getAsync<_i743.GetIdentityUseCase>(),
               gh<_i920.ProofGenerationStepsStreamManager>(),
               gh<_i267.StacktraceManager>(),
-              await getAsync<_i871.GetAuthTokenUseCase>(),
-              gh<_i698.Iden3commCredentialRepository>(),
-              gh<_i958.RemoveClaimsUseCase>(),
-              gh<_i635.SaveClaimsUseCase>(),
               await getAsync<_i143.RefreshCredentialUseCase>(),
+            ));
+    gh.factoryAsync<_i279.AddNewIdentityUseCase>(
+        () async => _i279.AddNewIdentityUseCase(
+              await getAsync<_i26.IdentityRepository>(),
+              await getAsync<_i561.AddIdentityUseCase>(),
+              gh<_i267.StacktraceManager>(),
             ));
     gh.factoryAsync<_i17.Identity>(() async => _i17.Identity(
           await getAsync<_i548.CheckIdentityValidityUseCase>(),
@@ -1345,16 +1185,9 @@ extension GetItInjectableX on _i174.GetIt {
           await getAsync<_i1050.AddProfileUseCase>(),
           await getAsync<_i561.GetProfilesUseCase>(),
           await getAsync<_i829.RemoveProfileUseCase>(),
-          await getAsync<_i78.GetDidUseCase>(),
+          gh<_i78.GetDidUseCase>(),
           gh<_i267.StacktraceManager>(),
         ));
-    gh.factoryAsync<_i989.UpdateInteractionUseCase>(
-        () async => _i989.UpdateInteractionUseCase(
-              await getAsync<_i1012.InteractionRepository>(),
-              gh<_i192.CheckProfileValidityUseCase>(),
-              await getAsync<_i743.GetIdentityUseCase>(),
-              await getAsync<_i1031.AddInteractionUseCase>(),
-            ));
     gh.factoryAsync<_i411.AuthenticateUseCase>(
         () async => _i411.AuthenticateUseCase(
               gh<_i88.Iden3commRepository>(),
@@ -1403,8 +1236,6 @@ class _$ChannelModule extends _i335.ChannelModule {}
 class _$NetworkModule extends _i335.NetworkModule {}
 
 class _$FilesManagerModule extends _i335.FilesManagerModule {}
-
-class _$KMSModule extends _i335.KMSModule {}
 
 class _$PlatformModule extends _i335.PlatformModule {}
 
