@@ -1,7 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';
-import 'package:polygonid_flutter_sdk/common/domain/use_cases/get_env_use_case.dart';
 import 'package:polygonid_flutter_sdk/common/infrastructure/stacktrace_stream_manager.dart';
 import 'package:polygonid_flutter_sdk/credential/domain/repositories/credential_repository.dart';
 import 'package:polygonid_flutter_sdk/iden3comm/domain/exceptions/iden3comm_exceptions.dart';
@@ -9,7 +8,6 @@ import 'package:polygonid_flutter_sdk/iden3comm/domain/repositories/iden3comm_re
 import 'package:polygonid_flutter_sdk/iden3comm/domain/use_cases/get_auth_inputs_use_case.dart';
 import 'package:polygonid_flutter_sdk/identity/domain/repositories/identity_repository.dart';
 import 'package:polygonid_flutter_sdk/identity/domain/repositories/smt_repository.dart';
-import 'package:polygonid_flutter_sdk/identity/domain/use_cases/get_did_identifier_use_case.dart';
 import 'package:polygonid_flutter_sdk/identity/domain/use_cases/get_latest_state_use_case.dart';
 import 'package:polygonid_flutter_sdk/identity/domain/use_cases/identity/get_identity_use_case.dart';
 import 'package:polygonid_flutter_sdk/identity/domain/use_cases/identity/sign_message_use_case.dart';
@@ -84,7 +82,8 @@ void main() {
             nonRevProof: anyNamed('nonRevProof'),
             gistProof: anyNamed('gistProof'),
             treeState: anyNamed('treeState')))
-        .thenAnswer((realInvocation) => Future.value(CommonMocks.aBytes));
+        .thenAnswer((realInvocation) =>
+            Future.value(CommonMocks.generateInputsResponse));
     when(identityRepository.getAuthClaimNode(children: anyNamed('children')))
         .thenAnswer((realInvocation) => Future.value(IdentityMocks.node));
     when(smtRepository.generateProof(
@@ -103,7 +102,8 @@ void main() {
     'Given a param, when I call execute, then I expect a bytes list to be returned',
     () async {
       // When
-      expect(await useCase.execute(param: param), CommonMocks.aBytes);
+      expect(await useCase.execute(param: param),
+          CommonMocks.generateInputsResponse);
 
       // Then
       var capturedIdentity =
