@@ -59,8 +59,8 @@ void main() {
         (realInvocation) => Future.value(Iden3commMocks.encodedJWZ));
     when(getAuthChallengeUseCase.execute(param: anyNamed('param')))
         .thenAnswer((realInvocation) => Future.value(CommonMocks.challenge));
-    when(getAuthInputsUseCase.execute(param: anyNamed('param')))
-        .thenAnswer((realInvocation) => Future.value(CommonMocks.aBytes));
+    when(getAuthInputsUseCase.execute(param: anyNamed('param'))).thenAnswer(
+        (realInvocation) => Future.value(CommonMocks.generateInputsResponse));
     when(loadCircuitUseCase.execute(param: anyNamed('param')))
         .thenAnswer((realInvocation) => Future.value(ProofMocks.circuitData));
     when(proveUseCase.execute(param: anyNamed('param')))
@@ -105,7 +105,7 @@ void main() {
         verify(proveUseCase.execute(param: captureAnyNamed('param')))
             .captured
             .first;
-    expect(captureProve.inputs, CommonMocks.aBytes);
+    expect(captureProve.inputs, CommonMocks.inputs);
     expect(captureProve.circuitData, ProofMocks.circuitData);
   });
 
@@ -150,7 +150,7 @@ void main() {
         verify(proveUseCase.execute(param: captureAnyNamed('param')))
             .captured
             .first;
-    expect(captureProve.inputs, CommonMocks.aBytes);
+    expect(captureProve.inputs, CommonMocks.inputs);
     expect(captureProve.circuitData, ProofMocks.circuitData);
   });
 }
