@@ -20,9 +20,6 @@ import 'package:polygonid_flutter_sdk/identity/data/mappers/private_key/private_
 import 'package:polygonid_flutter_sdk/identity/domain/entities/hash_entity.dart';
 import 'package:polygonid_flutter_sdk/identity/domain/entities/identity_entity.dart';
 import 'package:polygonid_flutter_sdk/identity/domain/entities/node_entity.dart';
-import 'package:polygonid_flutter_sdk/identity/data/dtos/rhs_node_dto.dart';
-import 'package:polygonid_flutter_sdk/identity/data/mappers/hex_mapper.dart';
-import 'package:polygonid_flutter_sdk/identity/data/mappers/rhs_node_mapper.dart';
 import 'package:polygonid_flutter_sdk/identity/data/mappers/state_identifier_mapper.dart';
 import 'package:polygonid_flutter_sdk/identity/data/repositories/identity_repository_impl.dart';
 import 'package:polygonid_flutter_sdk/identity/domain/entities/rhs_node_entity.dart';
@@ -68,14 +65,14 @@ final mockAuthResponse = AuthResponse(
 );
 
 final rhsNodeDTOs = [
-  RhsNodeDTO(
+  RhsNodeEntity(
       node: NodeEntity(
         children: const [],
         hash: HashEntity(data: Uint8List(32)),
         type: NodeType.middle,
       ),
       status: ''),
-  RhsNodeDTO(
+  RhsNodeEntity(
       node: NodeEntity(
         children: const [],
         hash: HashEntity(data: Uint8List(32)),
@@ -85,15 +82,19 @@ final rhsNodeDTOs = [
 ];
 final rhsNodeEntities = [
   RhsNodeEntity(
-    node: {},
-    status: '',
-    nodeType: RhsNodeType.state,
-  ),
+      node: NodeEntity(
+        children: const [],
+        hash: HashEntity(data: Uint8List(32)),
+        type: NodeType.middle,
+      ),
+      status: ''),
   RhsNodeEntity(
-    node: {},
-    status: '',
-    nodeType: RhsNodeType.state,
-  ),
+      node: NodeEntity(
+        children: const [],
+        hash: HashEntity(data: Uint8List(32)),
+        type: NodeType.leaf,
+      ),
+      status: ''),
 ];
 
 Response errorResponse = Response("body", 450);
@@ -151,9 +152,7 @@ MockEncryptionDbDataSource encryptionDbDataSource =
     MockEncryptionDbDataSource();
 MockDestinationPathDataSource destinationPathDataSource =
     MockDestinationPathDataSource();
-MockHexMapper hexMapper = MockHexMapper();
 MockPrivateKeyMapper privateKeyMapper = MockPrivateKeyMapper();
-MockRhsNodeMapper rhsNodeMapper = MockRhsNodeMapper();
 MockStateIdentifierMapper stateIdentifierMapper = MockStateIdentifierMapper();
 MockSecureStorageProfilesDataSource secureStorageProfilesDataSource =
     MockSecureStorageProfilesDataSource();
@@ -168,9 +167,7 @@ IdentityRepository repository = IdentityRepositoryImpl(
   libPolygonIdCoreIdentityDataSource,
   encryptionDbDataSource,
   destinationPathDataSource,
-  hexMapper,
   privateKeyMapper,
-  rhsNodeMapper,
   stateIdentifierMapper,
   secureStorageProfilesDataSource,
 );
@@ -185,9 +182,7 @@ IdentityRepository repository = IdentityRepositoryImpl(
   LibPolygonIdCoreIdentityDataSource,
   EncryptionDbDataSource,
   DestinationPathDataSource,
-  HexMapper,
   PrivateKeyMapper,
-  RhsNodeMapper,
   StateIdentifierMapper,
   SecureStorageProfilesDataSource,
 ])
@@ -201,8 +196,6 @@ void main() {
       when(walletDataSource.signMessage(
               privateKey: anyNamed('privateKey'), message: anyNamed('message')))
           .thenAnswer((realInvocation) => Future.value(CommonMocks.signature));
-      when(hexMapper.mapTo(any))
-          .thenAnswer((realInvocation) => CommonMocks.aBytes);
     });
 
     test(
@@ -218,13 +211,11 @@ void main() {
       );
 
       // Then
-      expect(verify(hexMapper.mapTo(captureAny)).captured.first,
-          CommonMocks.privateKey);
       var signCaptured = verify(walletDataSource.signMessage(
               privateKey: captureAnyNamed('privateKey'),
               message: captureAnyNamed('message')))
           .captured;
-      expect(signCaptured[0], CommonMocks.aBytes);
+      expect(signCaptured[0], [0, 0, 0]);
       expect(signCaptured[1], CommonMocks.message);
     });
 
@@ -247,13 +238,11 @@ void main() {
       });
 
       // Then
-      expect(verify(hexMapper.mapTo(captureAny)).captured.first,
-          CommonMocks.privateKey);
       var signCaptured = verify(walletDataSource.signMessage(
               privateKey: captureAnyNamed('privateKey'),
               message: captureAnyNamed('message')))
           .captured;
-      expect(signCaptured[0], CommonMocks.aBytes);
+      expect(signCaptured[0], [0, 0, 0]);
       expect(signCaptured[1], CommonMocks.message);
     });
   });
@@ -373,57 +362,10 @@ void main() {
   group("Get State Roots", () {
     setUp(() {
       reset(remoteIdentityDataSource);
-      reset(rhsNodeMapper);
 
       // Given
       when(remoteIdentityDataSource.fetchStateRoots(url: anyNamed('url')))
           .thenAnswer((realInvocation) => Future.value(rhsNodeDTOs[0]));
-      when(rhsNodeMapper.mapFrom(any)).thenReturn(rhsNodeEntities[0]);
-    });
-
-    test(
-        "Given parameters, when I call fetchStateRoots, then I expect a RhsNodeEntity to be returned",
-        () async {
-      // When
-      expect(await repository.getStateRoots(url: CommonMocks.url),
-          rhsNodeEntities[0]);
-
-      // Then
-      var fetchCaptured = verify(remoteIdentityDataSource.fetchStateRoots(
-              url: captureAnyNamed('url')))
-          .captured;
-
-      expect(fetchCaptured[0], CommonMocks.url);
-
-      expect(verify(rhsNodeMapper.mapFrom(captureAny)).captured.first,
-          rhsNodeDTOs[0]);
-    });
-
-    test(
-        "Given parameters, when I call fetchStateRoots and an error occurred, then I expect a FetchIdentityStateException to be thrown",
-        () async {
-      // Given
-      when(remoteIdentityDataSource.fetchStateRoots(
-        url: anyNamed('url'),
-      )).thenAnswer((realInvocation) => Future.error(CommonMocks.exception));
-
-      // When
-      await repository
-          .getStateRoots(url: CommonMocks.url)
-          .then((_) => expect(true, false))
-          .catchError((error) {
-        expect(error, isA<FetchStateRootsException>());
-        expect(error.error, CommonMocks.exception);
-      });
-
-      // Then
-      var fetchCaptured = verify(remoteIdentityDataSource.fetchStateRoots(
-        url: captureAnyNamed('url'),
-      )).captured;
-
-      expect(fetchCaptured[0], CommonMocks.url);
-
-      verifyNever(rhsNodeMapper.mapFrom(captureAny));
     });
   });
 
